+++ RJreferences.bib
@@ -1,39 +1,4 @@
 % put bibtex entries here!
-@Book{plotly,
-    author = {Carson Sievert},
-    title = {{Interactive Web-Based Data Visualizatio}n with R, plotly, and shiny},
-    publisher = {Chapman and Hall/CRC},
-    year = {2020},
-    isbn = {9781138331457},
-    url = {https://plotly-r.com},
-  }
-
-@Manual{crosstalk,
-  title = {{crosstalk}: Inter-Widget Interactivity for HTML Widgets},
-  author = {Joe Cheng and Carson Sievert},
-  year = {2021},
-  note = {R package version 1.1.1},
-  url = {https://CRAN.R-project.org/package=crosstalk},
-}
-
-@article{RJ-2021-050,
-  author = {Earo Wang and Dianne Cook},
-  title = {Conversations in time: interactive visualisation to explore
-          structured temporal data},
-  year = {2021},
-  journal = {The R Journal},
-  doi = {10.32614/RJ-2021-050},
-  url = {https://journal.r-project.org/archive/2021/RJ-2021-050/index.html}
-}
-
-@Manual{palmerpenguins,
-  title = {{palmerpenguins}: Palmer Archipelago (Antarctica) penguin data},
-  author = {Allison Marie Horst and Alison Presmanes Hill and Kristen B Gorman},
-  year = {2020},
-  note = {R package version 0.1.0},
-  url = {https://allisonhorst.github.io/palmerpenguins/},
-}
-
 
 @misc{22arro,
 	title = {22  Arrow {\textendash} R for Data Science (2e)},
@@ -102,6 +67,7 @@ PMCID: PMC6331976},
 	author = {Dolladille, Charles and {Chrétien}, Basile},
 	year = {2025},
 	date = {2025},
+	doi = {10.32614/cran.package.vigicaen},
 	url = {https://github.com/pharmacologie-caen/vigicaen},
 	langid = {en}
 }
@@ -121,5 +87,6 @@ PMCID: PMC6331976},
 	year = {2023},
 	month = {02},
 	date = {2023-02-22},
+	doi = {10.32614/cran.package.tidyverse},
 	url = {https://cran.r-project.org/web/packages/tidyverse/index.html}
 }
