+++ README.md
@@ -206,5 +206,12 @@ demo |>
 #> 1 a_colitis nivolumab    44  31.2 1.88  (1.23-2.88) 0.489  0.0314 95%
 ```
 
+## Code of Conduct
+
+Please note that the vigicaen project is released with a [Contributor
+Code of
+Conduct](https://pharmacologie-caen.github.io/vigicaen/CODE_OF_CONDUCT.html).
+By contributing to this project, you agree to abide by its terms.
+
 <!-- Footnote for myself
 &#10;You'll still need to render `README.Rmd` regularly, to keep `README.md` up-to-date. `devtools::build_readme()` is handy for this. You could also use GitHub Actions to re-render `README.Rmd` every time you push. An example workflow can be found here: <https://github.com/r-lib/actions/tree/v1/examples>. -->
