### 2025-09-16 ad620bf
Deploying to gh-pages from @ pharmacologie-caen/vigicaen@895543849679cf28d2d49fe03db7b464a64a4903 🚀
