### 2025-09-18 3ed773a
remove dependency to fst
