+++ DESCRIPTION
@@ -41,7 +41,6 @@ Imports:
     cli,
     dplyr,
     data.table,
-    fst,
     ggplot2,
     glue,
     gridExtra,
