+++ pkgdown.yml
@@ -10,7 +10,7 @@ articles:
   template_dictionary: template_dictionary.html
   template_main: template_main.html
   template_routine: template_routine.html
-last_built: 2025-09-16T09:28Z
+last_built: 2025-09-18T08:04Z
 urls:
   reference: https://pharmacologie-caen.github.io/vigicaen/reference
   article: https://pharmacologie-caen.github.io/vigicaen/articles
