### 2025-09-19 3ba50ea
patch for low spec computer and start of doc update (#158)
