+++ R/vigi_routine.R
@@ -65,7 +65,7 @@
 #'
 #' # But you could also use get_drecno() and get_llt_soc()
 #'
-#' # load tables demo, drug, adr, and link
+#' # load tables demo, drug, adr, and link (real tables with open_dataset())
 #'
 #' demo <- demo_
 #' adr  <- adr_
@@ -132,6 +132,10 @@ vigi_routine <-
     suspect_only = FALSE,
     d_code_2
   ){
+    # #### 0. arrow options #### ####
+
+    options(arrow.pull_as_vector = FALSE)
+
     # #### 0. checkers #### ####
     check_id_list_numeric(d_code)
     check_id_list_numeric(a_code)
@@ -148,7 +152,7 @@ vigi_routine <-
         )
       }
     }
-    check_data_demo(demo_data)
+
     check_data_adr(adr_data)
     check_data_drug(drug_data)
     check_data_link(link_data)
@@ -165,6 +169,14 @@ vigi_routine <-
     repbasis_sel <-
       if (suspect_only) {"s"} else {"sci"}
 
+    basis_sel <-
+      c(
+        if(grepl("s", repbasis_sel)){ 1 },
+        # subsidiary_files / Repbasis_Lx
+        if(grepl("c", repbasis_sel)){ 2 },
+        if(grepl("i", repbasis_sel)){ 3 }
+      )
+
     use_two_drugs <-
       !rlang::is_missing(d_code_2)
 
@@ -174,42 +186,63 @@ vigi_routine <-
       check_length_one(d_code_2, "vigi_routine()")
       d_name2 <- names(d_code_2)
 
-      # Demo data management with 2 drugs
-      suppressMessages(
-        demo_data <- demo_data |>
-          add_drug(d_code, drug_data = drug_data, repbasis = repbasis_sel) |>
-          add_drug(d_code_2, drug_data = drug_data, repbasis = repbasis_sel) |>
-          add_adr(a_code, adr_data = adr_data)
-      )
-      # Variable both_drugs to replace d_name, if d_code_2 provided
-      demo_data <-
-        demo_data |>
-        dplyr::mutate(both_drugs = ifelse(.data[[d_name]] == 1 &
-                                            .data[[d_name2]] == 1, 1, 0))
+      umc_drug_1 <-
+        drug_data |>
+        dplyr::filter(.data$DrecNo %in% d_code[[1]] &
+                        .data$Basis %in% basis_sel) |>
+        dplyr::pull(.data$UMCReportId) |>
+        unique()
+
+      umc_drug_2 <-
+        drug_data |>
+        dplyr::filter(.data$DrecNo %in% d_code_2[[1]] &
+                        .data$Basis %in% basis_sel) |>
+        dplyr::pull(.data$UMCReportId) |>
+        unique()
+
+      umc_drug <- intersect(umc_drug_1, umc_drug_2)
+
+      umc_adr <-
+        adr_data |>
+        dplyr::filter(.data$MedDRA_Id %in% a_code[[1]])  |>
+        dplyr::pull(.data$UMCReportId) |>
+        unique()
+
+
+      umc_cases <- intersect(umc_drug, umc_adr)
+
       cli::cli_alert_info(
         "Dual drug analysis: only cases exposed to both '{d_name}' and '{d_name2}' are included.")
 
       d_name_after_dm <- "both_drugs"
 
     } else {
-      # else, demo data management with 1 drug
-      suppressMessages(
-        demo_data <- demo_data |>
-          add_drug(d_code, drug_data = drug_data, repbasis = repbasis_sel) |>
-          add_adr(a_code, adr_data = adr_data)
-      )
+
+      umc_drug <-
+        drug_data |>
[…]
