+++ vignettes/template_routine.Rmd
@@ -31,12 +31,12 @@ path_meddra <- "~/meddra/"
 
 # ---- Load data ---- ####
 
-demo <- dt_parquet(path_base, "demo", in_memory = FALSE)
-drug <- dt_parquet(path_base, "drug", in_memory = FALSE)
-adr  <- dt_parquet(path_base, "adr",  in_memory = FALSE)
-link <- dt_parquet(path_base, "link", in_memory = FALSE)
+demo <- open_dataset(paste0(path_base, "demo.parquet"))
+drug <- open_dataset(paste0(path_base, "drug.parquet"))
+adr  <- open_dataset(paste0(path_base, "adr.parquet" ))
+link <- open_dataset(paste0(path_base, "link.parquet"))
 
-mp <- dt_parquet(path_who,  "mp")
+mp       <- dt_parquet(path_who,    "mp")
 meddra   <- dt_parquet(path_meddra, "meddra")
 
 # ---- Select drug and adverse drug reaction ---- ####
