### 2025-10-24 58adc71
handling original user option for arrow and using back dt_parquet in doc
