+++ vignettes/template_routine.Rmd
@@ -31,10 +31,10 @@ path_meddra <- "~/meddra/"
 
 # ---- Load data ---- ####
 
-demo <- open_dataset(paste0(path_base, "demo.parquet"))
-drug <- open_dataset(paste0(path_base, "drug.parquet"))
-adr  <- open_dataset(paste0(path_base, "adr.parquet" ))
-link <- open_dataset(paste0(path_base, "link.parquet"))
+demo <- dt_parquet(path_base, "demo", in_memory = FALSE)
+drug <- dt_parquet(path_base, "drug", in_memory = FALSE)
+adr  <- dt_parquet(path_base, "adr",  in_memory = FALSE)
+link <- dt_parquet(path_base, "link", in_memory = FALSE)
 
 mp       <- dt_parquet(path_who,    "mp")
 meddra   <- dt_parquet(path_meddra, "meddra")
