### 2025-10-30 0865a99
typo and doc update
