### 2025-10-30 881cb04
fixing a potential issue for the snapshot tests on github
