### 2025-10-30 5861945
first version of add_ind and test
