+++ R/add_ind.R
@@ -1,0 +1,215 @@
+#' Add indication
+#'
+#' @description `r lifecycle::badge('experimental')` Creates indication columns.
+#' in vigibase datasets (demo, link, adr, drug, or ind).
+#'
+#' @details Indication terms are issued from either MedDRA or International
+#' Classification of Diseases (ICD) - you need to use *both* dictionaries, should
+#' you wish to capture all terms related to a specific disease.
+#' Indication terms are not translated into codes in VigiBase ECL,
+#' unlike drug or adr terms. Therefore, there is no `get_*` step to collect
+#' such codes. The terms are passed directly to `i_list`, which should still be
+#' a *named* list containing indication terms.
+#'
+#' @param .data   The dataset used to identify individual reports (usually, it is `demo`)
+#' @param i_list  A named list of indication terms. See Details.
+#' @param i_names A character vector. Names for indication columns (must be the same length as i_list), default to `names(i_list)`
+#' @param drug_data A data.frame containing the drug data (usually, it is `drug`)
+#' @param ind_data  A data.frame containing the indication data (usually, it is `ind`)
+#'
+#' @returns A dataset with the new indication columns.
+#' Each element of `i_names` will add a column with the same name in `.data`.
+#' The value can be
+#' \itemize{
+#'  \item 0 The corresponding indication is absent.
+#'  \item 1 The indication is present in the case if `.data` is `demo` or `adr`,
+#' or "this row correspond to this indication",
+#' if `.data` is `drug`, `link` or `ind`).
+#'  \item NA There is no indication data for this case / drug.
+#'  }
+#' @export
+#'
+#' @importFrom rlang .data
+#' @importFrom rlang .env
+#' @importFrom dplyr if_else
+#' @seealso [add_adr()], [add_drug()]
+#'
+#' @examples
+#'
+#' # Set up a list of indication terms
+#'
+#' i_list <-
+#'   list(
+#'     melanoma = c("Malignant melanoma", "Metastatic malignant melanoma"),
+#'     lung_cancer = c("Non-small cell lung cancer", "Lung adenocarcinoma")
+#'     )
+#'
+#'  demo <-
+#'    demo_ |>
+#'    add_ind(i_list,
+#'            drug_data = drug_,
+#'            ind_data  = ind_)
+#'
+#'  demo |> desc_facvar(names(i_list))
+
+add_ind <-
+  function(.data,
+           i_list,
+           i_names = names(i_list),
+           drug_data,
+           ind_data){
+
+    # 0. arrow options
+
+    # keep original user option, then set it
+
+    original_user_option <- options("arrow.pull_as_vector")
+
+    options(arrow.pull_as_vector = FALSE)
+
+    # 1. checkers
+
+    check_data_drug(drug_data, "drug_data")
+    check_data_ind(ind_data, "ind_data")
+
+    data_type <-
+      query_data_type(.data, ".data")
+
+    # 2. identify table_ids to collect
+
+    t_id <-
+      switch(data_type,
+             demo = "UMCReportId",
+             adr  = "UMCReportId",
+             link = "Drug_Id",
+             drug = "Drug_Id",
+             ind  = "Drug_Id"
+      )
+
+    renamer_tid <-
+      c("t_id" = t_id)
+
+    dd <-
+      drug_data |>
+      dplyr::rename(dplyr::all_of(renamer_tid))
+
+    # 3. collect drug_ids from ind_data
+
+    d_ids <-
+      purrr::map(i_list, function(i_list_batch) {
+        ind_data |>
[…]
