### 2025-10-31 70c1c84
Deploying to gh-pages from @ pharmacologie-caen/vigicaen@f573f2f2fc9cd3e870094a24c7d7a20c9e69f396 🚀
