+++ pkgdown.yml
@@ -10,7 +10,7 @@ articles:
   template_dictionary: template_dictionary.html
   template_main: template_main.html
   template_routine: template_routine.html
-last_built: 2025-09-18T08:04Z
+last_built: 2025-10-31T11:12Z
 urls:
   reference: https://pharmacologie-caen.github.io/vigicaen/reference
   article: https://pharmacologie-caen.github.io/vigicaen/articles
