+++ vignettes/template_dictionary.Rmd
@@ -142,5 +142,13 @@ a_llt <-
                 smq_list,
                 smq_content)
   )
+
+# ---- Set indication terms ---- ####
+
+i_list <-
+ list(
+   melanoma = c("Malignant melanoma", "Metastatic malignant melanoma"),
+   lung_cancer = c("Non-small cell lung cancer", "Lung adenocarcinoma")
+   )
 ```
 
