+++ vignettes/template_main.Rmd
@@ -80,6 +80,7 @@ link     <- link_
 out      <- out_
 followup <- followup_
 srce     <- srce_
+ind      <- ind_
 
 thg      <- thg_
 mp       <- mp_
@@ -104,6 +105,12 @@ atc_drecno <-
                vigilyze = TRUE)
 
 a_llt <- ex_$a_llt
+
+i_list <-
+ list(
+   melanoma = c("Malignant melanoma", "Metastatic malignant melanoma"),
+   lung_cancer = c("Non-small cell lung cancer", "Lung adenocarcinoma")
+   )
 ```
 
 ## Demo
@@ -143,6 +150,16 @@ demo <-
     adr_data = adr
   )
 
+# ---- Indications ---- ####
+
+demo <-
+  demo |>
+  add_ind(
+    i_list,
+    drug_data = drug,
+    ind_data = ind
+  )
+
 # ---- Demographics ---- ####
 
 # Age, sex
@@ -180,7 +197,7 @@ demo <-
 demo <-
   demo |>
   mutate(
-    fup = if_else(UMCReportId %in% followup$UMCReportId, 1, 0),
+    fup = ifelse(UMCReportId %in% followup$UMCReportId, 1, 0),
     serious = 
       ifelse(
         UMCReportId %in% out$UMCReportId,
