+++ man/tb_vigibase.Rd
@@ -64,7 +64,7 @@ If set to FALSE (the default), the function will skip the construction of any .p
 so you do not have to start from scratch after a failure. Set to TRUE to force rebuilding all tables.
 }
 \examples{
-\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
+\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
 
 # --- Set up example source files ---- ####
 
