+++ DESCRIPTION
@@ -25,7 +25,7 @@ Encoding: UTF-8
 LazyData: true
 LazyDataCompression: xz
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.3.2
+RoxygenNote: 7.3.3
 URL: https://github.com/pharmacologie-caen/vigicaen, https://pharmacologie-caen.github.io/vigicaen/
 BugReports: https://github.com/pharmacologie-caen/vigicaen/issues
 Suggests: 
