### 2025-11-03 4282990
wip on finalizing add_dose and its test
