+++ tests/testthat/test-add_dose.R
@@ -1,22 +1,50 @@
-library(testthat)
-library(dplyr)
-library(purrr)
+test_that("finding or not finding drug dose displays correctly", {
+  d_code <- list(paracetamol = c(97818920, 97409107),
+                 unknown_drug = c(99999999))
 
-# Dummy Data Setup
-demo <- demo_
-drug <- drug_
+  expect_snapshot(
+      demo <-
+        add_dose(
+        .data = demo_,
+        d_code = d_code[1],
+        drug_data = drug_
+      )
+  )
 
-# Unit Tests
+  # with no dose
+  expect_snapshot(
+    demo <-
+      add_dose(
+        .data = demo_,
+        d_code = d_code[2],
+        drug_data = drug_
+      )
+  )
+
+  # both
+
+  expect_snapshot(
+    demo <-
+      add_dose(
+        .data = demo_,
+        d_code = d_code,
+        drug_data = drug_
+      )
+  )
+})
+
 test_that("add_dose correctly adds daily dose columns", {
   d_code <- list(paracetamol = c(97818920, 97409107))
 
-  result <- add_dose(
-    .data = demo,
-    d_code = d_code,
-    d_names = names(d_code),
-    repbasis = "sci",
-    method = "DrecNo",
-    drug_data = drug
+  suppressMessages(
+    result <- add_dose(
+      .data = demo_,
+      d_code = d_code,
+      d_dose_names = names(d_code),
+      repbasis = "sci",
+      method = "DrecNo",
+      drug_data = drug_
+    )
   )
 
   expect_true("paracetamol_dose_mg_per_day" %in% colnames(result))
@@ -29,13 +57,15 @@ test_that("works with irregular drug and demo names", {
   dema <- demo_
   druga <- drug_
 
-  result <- add_dose(
-    .data = dema,
-    d_code = d_code,
-    d_names = names(d_code),
-    repbasis = "sci",
-    method = "DrecNo",
-    drug_data = druga
+  suppressMessages(
+    result <- add_dose(
+      .data = dema,
+      d_code = d_code,
+      d_dose_names = names(d_code),
+      repbasis = "sci",
+      method = "DrecNo",
+      drug_data = druga
+    )
   )
 
   expect_true("paracetamol_dose_mg_per_day" %in% colnames(result))
@@ -56,13 +86,15 @@ test_that("add_dose handles empty dataset gracefully", {
 
   d_code <- list(paracetamol = c(97818920, 97409107))
 
-  result <- add_dose(
-    .data = empty_demo,
-    d_code = d_code,
-    d_names = names(d_code),
-    repbasis = "sci",
-    method = "DrecNo",
-    drug_data = drug
+  suppressMessages(
+    result <- add_dose(
+      .data = empty_demo,
+      d_code = d_code,
+      d_dose_names = names(d_code),
+      repbasis = "sci",
+      method = "DrecNo",
+      drug_data = drug_
+    )
   )
 
   expect_equal(nrow(result), 0)  # Should return an empty dataset
@@ -73,12 +105,11 @@ test_that("add_dose correctly filters and calculates doses", {
   d_code <- list(paracetamol = c(97818920, 97409107))
 
   result <- add_dose(
-    .data = demo,
+    .data = demo_,
     d_code = d_code,
-    d_names = names(d_code),
     repbasis = "sci",
     method = "DrecNo",
-    drug_data = drug
+    drug_data = drug_
   )
 
   # Check if the dose is calculated correctly (based on the example in the function)
@@ -89,21 +120,21 @@ test_that("add_dose works for different `repbasis` values", {
[…]
