+++ R/add_dose.R
@@ -1,34 +1,36 @@
 #' Add Dose in mg to a dataset
 #'
 #' @description `r lifecycle::badge('experimental')`
-#' `add_dose()` creates dynamic columns representing drug doses (in mg)
+#' `add_dose()` creates drug dose columns in vigibase
+#' datasets (demo, link, adr, drug, ind)
 #' for specified drugs in a dataset. It calculates daily dose values
 #' based on dose amount, frequency, and their corresponding units.
-#' The function is #' compatible with `demo`, `link`, and `adr` datasets.
+#' The function is compatible with `demo`, `link`, `adr`, `drug` and `ind`
+#'  datasets.
 #'
 #' @details
+#' Currently, only drug doses in **mg per day** are handled.
 #' The function identifies drug doses in a dataset by cross-referencing
-#' with a drug data table. Drug codes can be specified using either
-#' DrecNos or MedicinalProd_Id. Doses are filtered based on reputation
+#' with a drug data table. Drugs may be filtered based on reputation
 #' bases (suspect, concomitant, or interacting).
-#' `d_code` is a named list containing drug codes.
 #' Either drug record numbers (e.g., from [get_drecno()]), or
-#' medicinalprod_ids (e.g., from [get_atc_code()]). Default method is to DrecNos.
+#' medicinalprod_ids (e.g., from [get_atc_code()]) can be used to
+#' identify drugs. Default method is to DrecNos.
 #'
-#' **Important:** Ensure the dataset's structure aligns with the `data_type`
-#' argument to avoid errors.
 #'
 #' **Note:** It is very important to check the results obtained, as coding problems
-#' are very frequent for posology data. Some results might appear unbelievable due
-#' to these issues and should be carefully reviewed and trimmed accordingly.
+#' are very frequent for dose data, and some results might appear unreliable.
 #'
 #' @param .data The dataset used to identify individual reports (usually, it is `demo`)
 #' @param d_code A named list of drug codes (DrecNos or MPI). See Details.
-#' @param d_names A character vector. Names for drug columns (must be the same length as d_code), default to `names(d_code)`
+#' @param d_dose_names A character vector. Names for drug dose
+#' columns (must be the same length as d_code), default to `names(d_code)`.
+#' Will be followed by a fixed suffix "_dose_mg_per_day".
 #' @param repbasis Suspect, interacting and/or concomitant.
 #' Type initial of those you wish to select ("s" for suspect, "c" for concomitant
 #' and "i" for interacting ; default to all, e.g. "sci").
-#' @param method A character string. The type of drug code (DrecNo or MedicinalProd_Id). See details.
+#' @param method A character string.
+#' The type of drug code (DrecNo or MedicinalProd_Id). See details.
 #' @param drug_data A data.frame containing the drug data (usually, it is `drug`)
 #'
 #'
@@ -38,25 +40,28 @@
 #' @seealso [add_drug()], [get_drecno()], [get_atc_code()]
 #' @examples
 #' # Example: Adding doses for paracetamol
-#' d_code <- list(paracetamol = c(001, 002))
-#' demo_updated <- add_dose(
-#'   .data = demo_,
-#'   d_code = d_code,
-#'   d_names = "paracetamol",
-#'   repbasis = "sci",
-#'   method = "DrecNo",
-#'   drug_data = drug_
-#' )
+#' d_code <- list(paracetamol = c(97818920, 97409107))
+#' demo <-
+#'   add_dose(
+#'     .data = demo_,
+#'     d_code = d_code,
[…]

[…]
