### 2025-11-05 6a8f7be
Deploying to gh-pages from @ pharmacologie-caen/vigicaen@c4ad06b202f09d4b23338d649a533390110d02f3 🚀
