### 2025-11-05 d6a0c05
simplified syntax and wip on output
