+++ R/add_dose.R
@@ -93,11 +93,11 @@ add_dose <-
       paste0(d_dose_names, "_dose_mg_per_day")
 
 
-    dd_rb <- drug_data |>
-      dplyr::filter(.data$Basis %in% basis_sel)
+    # dd_rb <- drug_data |>
+    #   dplyr::filter(.data$Basis %in% basis_sel)
 
     renamer_did <- c("did_col" = method)
-    dd_rb <- dd_rb |> dplyr::rename(dplyr::all_of(renamer_did))
+    # dd_rb <- dd_rb |> dplyr::rename(dplyr::all_of(renamer_did))
 
     # identify table_ids to collect
 
@@ -111,84 +111,63 @@ add_dose <-
       )
 
     renamer_tid <- c("t_id" = t_id)
-    dd_rb <- dd_rb |> dplyr::rename(dplyr::all_of(renamer_tid))
 
-    # Collect table_ids with doses
-    t_ids <- purrr::map(d_code, function(d_code_batch) {
-      dd_rb |>
-        dplyr::filter(.data$did_col %in% d_code_batch) |>
-        dplyr::select(t_id, Amount, AmountU, Frequency, FrequencyU) |>
-        dplyr::mutate(
-          AmountU = gsub("\\s+", "", AmountU),
-          FrequencyU = gsub("\\s+", "", FrequencyU),
-          Amount = gsub("\\s+", "", Amount),
-          Frequency = gsub("\\s+", "", Frequency)
-        ) |>
-        dplyr::filter(Amount != "-") |>
-        dplyr::filter(trimws(AmountU) %in% c("1", "2", "3", "4", "5", "6")) |>
-        dplyr::filter(Frequency != "-") |>
-        dplyr::filter(trimws(FrequencyU) %in% c("801", "802", "803", "804", "805", "806")) |>
-        dplyr::filter(Frequency != 0) |>
-        dplyr::mutate(
-          Amount = as.numeric(Amount),
-          Frequency = as.numeric(Frequency),
-          multiplicator_amount = dplyr::case_when(
-            AmountU == "1" ~ 1000000,
-            AmountU == "2" ~ 1000,
-            AmountU == "3" ~ 1,
-            AmountU == "4" ~ 1 / 1000,
-            AmountU == "5" ~ 1 / 1000000,
-            AmountU == "6" ~ 1 / 1000000000,
-            TRUE ~ NA_real_
-          ),
-          multiplicator_frequency = dplyr::case_when(
-            FrequencyU == "806" ~ 1440,
-            FrequencyU == "805" ~ 24,
-            FrequencyU == "804" ~ 1,
-            FrequencyU == "803" ~ 1 / 7,
-            FrequencyU == "802" ~ 1 / 30,
-            FrequencyU == "801" ~ 1 / 365.25,
-            TRUE ~ NA_real_
-          ),
-          dose_mg_per_day = (Amount * multiplicator_amount * multiplicator_frequency * Frequency
-          )) |>
-        dplyr::filter(!is.na(dose_mg_per_day)) |>
-        dplyr::group_by(t_id) |>
-        dplyr::slice_max(dose_mg_per_day, with_ties = FALSE) |>
-        dplyr::ungroup()|>
-        dplyr::select(t_id, dose_mg_per_day) # Only keep relevant columns
-    })
+    back_renamer_tid <-
+      "t_id" |> rlang::set_names(t_id)
+    # dd_rb <- dd_rb |> dplyr::rename(dplyr::all_of(renamer_tid))
+
+    # Collect doses for each t_id
 
 
-    e_l <-
-      t_ids |>
-      purrr::map(function(t_id_subset) {
-        # Create a lookup vector for t_id to dose
[…]
