### 2025-11-06 28f0780
extended ind table support to add_drug, erroring with add_adr, doc
