+++ tests/testthat/test-add_dose.R
@@ -1,6 +1,7 @@
 test_that("finding or not finding drug dose displays correctly", {
   d_code <- list(paracetamol = c(97818920, 97409107),
-                 unknown_drug = c(99999999))
+                 unknown_drug = c(99999999),
+                 para2 = c(97818920, 97409107))
 
   expect_snapshot(
       demo <-
@@ -29,6 +30,40 @@ test_that("finding or not finding drug dose displays correctly", {
         .data = demo_,
         d_code = d_code,
         drug_data = drug_
+      )
+  )
+
+  # arbitrarily complex combination
+
+  expect_snapshot(
+    demo <-
+      add_dose(
+        .data = demo_,
+        d_code = d_code[c(1, 2, 3, 1, 2)],
+        drug_data = drug_
+      )
+  )
+
+  # with verbose to FALSE
+
+  expect_snapshot(
+    demo <-
+      add_dose(
+        .data = demo_,
+        d_code = d_code[1],
+        drug_data = drug_,
+        verbose = FALSE
+      )
+  )
+
+  expect_snapshot( # verbose is not affecting
+    # warning about no dose found
+    demo <-
+      add_dose(
+        .data = demo_,
+        d_code = d_code[2],
+        drug_data = drug_,
+        verbose = FALSE
       )
   )
 })
@@ -104,12 +139,14 @@ test_that("add_dose handles empty dataset gracefully", {
 test_that("add_dose correctly filters and calculates doses", {
   d_code <- list(paracetamol = c(97818920, 97409107))
 
-  result <- add_dose(
-    .data = demo_,
-    d_code = d_code,
-    repbasis = "sci",
-    method = "DrecNo",
-    drug_data = drug_
+  suppressMessages(
+    result <- add_dose(
+      .data = demo_,
+      d_code = d_code,
+      repbasis = "sci",
+      method = "DrecNo",
+      drug_data = drug_
+    )
   )
 
   # Check if the dose is calculated correctly (based on the example in the function)
@@ -119,22 +156,26 @@ test_that("add_dose correctly filters and calculates doses", {
 test_that("add_dose works for different `repbasis` values", {
   d_code <- list(paracetamol = c(97818920, 97409107))
 
-  result_suspect <- add_dose(
-    .data = demo_,
-    d_code = d_code,
-    d_dose_names = names(d_code),
-    repbasis = "s",
-    method = "DrecNo",
-    drug_data = drug_
+  suppressMessages(
+    result_suspect <- add_dose(
+      .data = demo_,
+      d_code = d_code,
+      d_dose_names = names(d_code),
+      repbasis = "s",
+      method = "DrecNo",
+      drug_data = drug_
+    )
   )
 
-  result_concomitant <- add_dose(
-    .data = demo_,
-    d_code = d_code,
-    d_dose_names = names(d_code),
-    repbasis = "c",
-    method = "DrecNo",
-    drug_data = drug_
+  suppressMessages(
+    result_concomitant <- add_dose(
+      .data = demo_,
+      d_code = d_code,
+      d_dose_names = names(d_code),
+      repbasis = "c",
+      method = "DrecNo",
+      drug_data = drug_
+    )
   )
 
   # The column names should differ
@@ -142,7 +183,7 @@ test_that("add_dose works for different `repbasis` values", {
[…]
