### 2025-11-07 ca5666c
compatibility with arrow
