+++ R/add_dose.R
@@ -1,4 +1,4 @@
-#' Add Dose in mg to a dataset
+#' Add drug dose in milligram per day
 #'
 #' @description `r lifecycle::badge('experimental')`
 #' `add_dose()` creates drug dose columns in vigibase
@@ -6,20 +6,30 @@
 #' for specified drugs in a dataset. It calculates daily dose values
 #' based on dose amount, frequency, and their corresponding units.
 #' The function is compatible with `demo`, `link`, `adr`, `drug` and `ind`
-#'  datasets.
+#' datasets.
 #'
 #' @details
-#' Currently, only drug doses in **mg per day** are handled.
+#' Actual supported dosage regimens are any combination of:
+#' \itemize{
+#'   \item Kilograms, grams, milligrams, micrograms, nanograms, or picograms
+#'   \item Per minute, hour, day, week, month, or year.
+#'   }
+#' Note that the result will be expressed in **milligrams per day**, whatever the
+#' aforementioned combination is. This may lead to very small or very large amounts
+#' in `drug_dose_mg_per_day` columns,
+#' depending on the actual dosage regimen.
 #' The function identifies drug doses in a dataset by cross-referencing
-#' with a drug data table. Drugs may be filtered based on reputation
+#' with a `drug_data` table.
+#' If either the amount unit (grams, etc.) *or* the frequency (days, etc.) is missing
+#' in `drug_data`, the corresponding row will be omitted.
+#' Drugs may be filtered based on reputation
 #' bases (suspect, concomitant, or interacting).
 #' Either drug record numbers (e.g., from [get_drecno()]), or
 #' medicinalprod_ids (e.g., from [get_atc_code()]) can be used to
 #' identify drugs. Default method is to DrecNos.
 #'
-#'
-#' **Note:** It is very important to check the results obtained, as coding problems
-#' are very frequent for dose data, and some results might appear unreliable.
+#' **It is very important to check the results**, as coding issues
+#' are common for dose data, and some results may seem unreliable.
 #'
 #' @param .data The dataset used to identify individual reports (usually, it is `demo`)
 #' @param d_code A named list of drug codes (DrecNos or MPI). See Details.
@@ -50,7 +60,7 @@
 #'     drug_data = drug_
 #'     )
 #'
-#'  desc_facvar(demo, "paracetamol_dose_mg_per_day")
+#' desc_cont(demo, "paracetamol_dose_mg_per_day")
 #'
 #' # Use only drug dose where paracetamol had a "suspect" reputation base.
 #' demo <-
@@ -62,7 +72,7 @@
 #'     drug_data = drug_
 #'   )
 #'
-#'   desc_facvar(demo, "para_susp_dose_mg_per_day")
+#' desc_cont(demo, "para_susp_dose_mg_per_day")
 
 
 ###########
@@ -77,6 +87,17 @@ add_dose <-
            verbose = TRUE
   )
   {
+
+    # 0. arrow options
+
+    # keep original user option, then set it
+
+    original_user_option <- options("arrow.pull_as_vector")
+
+    options(arrow.pull_as_vector = TRUE) # ! different from other add_* functions
+    # as purpose is different (only used for a summary table)
+
+    # 1. checkers
 
     check_id_list_numeric(d_code)
     method <- rlang::arg_match(method)
@@ -169,7 +190,10 @@ add_dose <-
       rlang::set_names() |>
       purrr::map( ~ {
         drug_col <- paste0( .x, "_dose_mg_per_day")
-        sum(!is.na(.data[[drug_col]]))
+        .data |>
[…]
