### 2025-11-07 2c45af0
Deploying to gh-pages from @ pharmacologie-caen/vigicaen@dea9e5e37fa8254db73ca0597cf4c1c05dcf2d30 🚀
