+++ articles/descriptive.md
@@ -1,0 +1,1039 @@
+# Descriptive
+
+## Introduction
+
+### Objectives
+
+- Compute descriptive statistics on a (vigibase) dataset
+
+- Understand the structure of the `link` table
+
+### Prerequisite
+
+- Be familiar with data management on drugs and adrs, with the `get_*`
+  and `add_*` functions (see
+  [`vignette("basic_workflow")`](https://pharmacologie-caen.github.io/vigicaen/articles/basic_workflow.md))
+
+## Demo data: counts of drugs, adrs, case characteristics
+
+### Step 0: Load packages
+
+``` r
+library(vigicaen)
+library(rlang)
+library(dplyr)
+```
+
+### Step 1: Load datasets and add drug and adr columns
+
+This vignette uses the preloaded datasets (and a spurious suspdup
+table).
+
+``` r
+demo     <- demo_
+adr      <- adr_
+drug     <- drug_
+link     <- link_
+out      <- out_
+followup <- followup_
+
+srce     <- srce_
+
+thg      <- thg_
+mp       <- mp_
+meddra   <- meddra_
+smq_list <- smq_list_
+smq_content <- smq_content_
+
+suspdup <- 
+  data.table::data.table(
+    UMCReportId = 1,
+    SuspectedduplicateReportId = NA
+  )
+```
+
+And preloaded drug and adr dictionaries.
+
+``` r
+d_drecno <- ex_$d_drecno
+
+a_llt <- ex_$a_llt
+```
+
+``` r
+demo <-
+  demo |>
+  add_drug(
+    d_code = d_drecno,
+    drug_data = drug
+  )
+#> ℹ `.data` detected as `demo` table.
+
+demo <-
+  demo |>
+  add_adr(
+    a_code = a_llt,
+    adr_data = adr
+  )
+#> ℹ `.data` detected as `demo` table.
+```
+
+As we aim to describe drug and adr counts, but also other variables
+(age, sex, type of reporter), they will be added too.
+
+You can still refer to
+
+``` r
+# Age, sex
+
+demo <-
+  demo |>
+  mutate(
+    age = cut(as.integer(AgeGroup),
+              breaks = c(0,4,5,6,7,8),
+              include.lowest = TRUE, right = TRUE,
+              labels = c("<18", "18-45","45-64", "65-74", "75+")),
+
+    sex = case_when(Gender == "1" ~ 1,
+                    Gender == "2" ~ 2,
+                    Gender %in% c("-","0","9") ~ NA_real_,
+                    TRUE ~ NA_real_)
+  )
+
+# Death + outcome availability
+
+demo <- 
+  demo |> 
+  mutate(death = 
+           ifelse(UMCReportId %in% out$UMCReportId,
+                  UMCReportId %in% 
+                    (out |> 
+                    filter(Seriousness == "1") |> 
+                    pull(UMCReportId)
+                    ),
+                  NA)
+         )
+
+# follow-up, seriousness
+
+demo <-
+  demo |>
+  mutate(
+    fup = ifelse(UMCReportId %in% followup$UMCReportId, 1, 0),
+    serious = 
+      ifelse(
+        UMCReportId %in% out$UMCReportId,
+        UMCReportId %in% 
+          (out |> 
+          filter(Serious == "Y") |> 
+          pull(UMCReportId)
+          ),
+        NA)
+  )
+
+# year
+
+demo <- 
+  demo |> 
+  mutate(
+    year = as.numeric(substr(FirstDateDatabase, start = 1, stop = 4))
+    )
+
+# type of reporter
+
+demo <-
+  demo |>
+  left_join(
+    srce |> transmute(UMCReportId, type_reporter = Type),
+    by = "UMCReportId")
+```
+
[…]
