+++ reference/vigicaen-package.md
@@ -1,0 +1,37 @@
+# vigicaen: 'VigiBase' Pharmacovigilance Database Toolbox
+
+Perform the analysis of the World Health Organization (WHO)
+Pharmacovigilance database 'VigiBase' (Extract Case Level version),
+<https://who-umc.org/> e.g., load data, perform data management,
+disproportionality analysis, and descriptive statistics. Intended for
+pharmacovigilance routine use or studies. This package is NOT supported
+nor reflect the opinion of the WHO, or the Uppsala Monitoring Centre.
+Disproportionality methods are described by Norén et al (2013)
+[doi:10.1177/0962280211403604](https://doi.org/10.1177/0962280211403604)
+.
+
+## See also
+
+Useful links:
+
+- <https://github.com/pharmacologie-caen/vigicaen>
+
+- <https://pharmacologie-caen.github.io/vigicaen/>
+
+- Report bugs at <https://github.com/pharmacologie-caen/vigicaen/issues>
+
+## Author
+
+**Maintainer**: Charles Dolladille <cdolladille@hotmail.com>
+([ORCID](https://orcid.org/0000-0003-0449-6261))
+
+Authors:
+
+- Basile Chrétien ([ORCID](https://orcid.org/0000-0002-7483-2489))
+
+Other contributors:
+
+- Universite de Caen Normandie (Caen, France) \[copyright holder\]
+
+- Unite de pharmaco-epidemiologie (Service de pharmacologie, Centre
+  Hospitalier Universitaire de Caen, Caen, France) \[copyright holder\]
