### 2025-11-07 6df6ffe
attempt to shut out inappropriate failure for some vigi_routine tests
