+++ tests/testthat/test-vigi_routine.R
@@ -2,7 +2,10 @@
 # VDIFFR_RUN_TESTS is explicitly set to "true", which should be the case only on
 # a GitHub Actions CI runner with stable version of R.
 
-if (requireNamespace("vdiffr", quietly = TRUE) && utils::packageVersion('testthat') >= '3.0.3' && utils::packageVersion('base') >= '4.4.1') {
+if (requireNamespace("vdiffr", quietly = TRUE) &&
+    utils::packageVersion('testthat') >= '3.0.3' &&
+    utils::packageVersion('base') >= '4.4.1' &&
+    Sys.info()["sysname"] != "Darwin") {
   expect_doppelganger <- vdiffr::expect_doppelganger
 } else {
   # If vdiffr is not available and visual tests are explicitly required, raise error.
