### 2025-11-07 0c94181
Deploying to gh-pages from @ pharmacologie-caen/vigicaen@6df6ffe9ec4bf9c4c93bf15d994a404eb2c29fc7 🚀
