### 2025-11-07 3fe59f9
improving doc consistency (mostly function titles)
