+++ R/get_llt_soc.R
@@ -1,4 +1,4 @@
-#' Extract low level terms from soc classification
+#' Get low level term codes from soc classification
 #'
 #' @description `r lifecycle::badge('stable')` Collect llt codes from a `meddra`
 #' data.table, given another term of the MedDRA SOC Hierarchy.
