### 2025-11-07 0fda9a8
Deploying to gh-pages from @ pharmacologie-caen/vigicaen@3fe59f9c611d9a8b86f3d274f45c312877405f15 🚀
