+++ reference/meddra_.md
@@ -1,4 +1,4 @@
-# Sample of Meddra.
+# Sample of MedDRA
 
 Anonymized data from MedDRA, used to illustrate the package examples and
 vignettes. You can find term codes related to colitis, pneumonitis,
