### 2025-11-08 7c13cf9
switch read_parquet to open_dataset in tb_subset
