### 2025-11-08 d39335c
Deploying to gh-pages from @ pharmacologie-caen/vigicaen@5b5aefb88c0ecba1e70d44c14fd7fcbd4db135ba 🚀
