### 2025-11-08 49c0950
referencing relevant issue in news
