### 2025-11-08 0afde88
Deploying to gh-pages from @ pharmacologie-caen/vigicaen@49c0950e1a9c6b25e5e31a58d9331f25e7a679d9 🚀
