### 2026-02-02 cf0c8fb
adding new sections per editor query
