+++ paper.md
@@ -1,35 +1,39 @@
 ---
-title: "Vigicaen: A `vigibase®' Pharmacovigilance Database Toolbox."
+title: 'Vigicaen: A `vigibase®` Pharmacovigilance Database Toolbox.'
 tags:
-  - R
-  - Pharmacovigilance
-  - Database
-  - Disproportionality analysis
-authors:  
-  - name: Charles Dolladille
-    orcid: 0000-0003-0449-6261
-    affiliation: 1
-    url: https://www.linkedin.com/in/charles-dolladille-30090b2b0/
-    email:  dolladille-c@chu-caen.fr
-  - name: Basile Chrétien
-    orcid: 0000-0002-7483-2489
-    affiliation: 2
-    url: https://www.linkedin.com/in/basile-chretien/
-    email: basile.chretien@outlook.com
-affiliations:
-  - name: University of Caen Normandy, Pharmacology Department, Centre Hospitalier Universitaire de Caen, Caen, France
-    index: 1
-  - name: University of Nagoya, Department of biostatistics, Nagoya University Hospital, Nagoya, Japan
-    index: 2
-date: "2025-08-25"
-bibliography: RJreferences.bib
-type: package
-output: 
-  pdf_document: 
+- R
+- Pharmacovigilance
+- Database
+- Disproportionality analysis
+date: "2026-02-02"
+output:
+  pdf_document:
     keep_md: true
   rmarkdown::html_vignette: default
-editor_options: 
-  markdown: 
+  html_document:
+    df_print: paged
+authors:
+- name: Charles Dolladille
+  orcid: "0000-0003-0449-6261"
+  affiliation: 1
+  url: "https://www.linkedin.com/in/charles-dolladille-30090b2b0/"
+  email: "dolladille-c@chu-caen.fr"
+- name: Basile Chrétien
+  orcid: "0000-0002-7483-2489"
+  affiliation: 2
+  url: "https://www.linkedin.com/in/basile-chretien/"
+  email: basile.chretien@outlook.com
+bibliography: RJreferences.bib
+type: package
+affiliations:
+- name: University of Caen Normandy, Pharmacology Department, Centre Hospitalier Universitaire
+    de Caen, Caen, France
+  index: 1
+- name: University of Nagoya, Department of biostatistics, Nagoya University Hospital,
+    Nagoya, Japan
+  index: 2
+editor_options:
+  markdown:
     wrap: 72
 ---
 
@@ -53,7 +57,7 @@ resolving of common issues related to drug and adverse event
 identification. The package is intended for pharmacovigilance
 practitioners, clinicians and researchers with or without advanced
 biostatistical skills. A graphical output can be produced for routine
-use, to support daily assessment of causality.
+use, to support daily assessment of drug liability.
 
 # Statement of need
 
@@ -62,8 +66,8 @@ domain of drug safety signal detection. Advanced methodologies are
 required to address common biases within pharmacovigilance databases.
 These analyses necessitate expertise in biostatistical software, such as
 R, which may present substantial challenges in terms of acquiring and
-maintaining the requisite skills—in addition to a solid understanding of
-pharmacovigilance principles and reporting systems.
+maintaining the requisite skills — in addition to a solid understanding
+of pharmacovigilance principles and reporting systems.
 
 For decades, the World Health Organization (WHO) has been collecting
 adverse drug reaction reports, called Individual Case Safety Reports
@@ -84,7 +88,7 @@ chance alone. Together with an incertitude margin, these estimators are
 used to raise safety signals on drugs.[@montastruc2011]
 
 The Uppsala Monitoring Centre grants access to VigiBase® to researchers,
-either academic or industrial, under a licence contract. The most
+either academics or industrials, under a licence contract. The most
[…]
