+++ RJreferences.bib
@@ -1,4 +1,39 @@
 % put bibtex entries here!
+@Book{plotly,
+    author = {Carson Sievert},
+    title = {{Interactive Web-Based Data Visualizatio}n with R, plotly, and shiny},
+    publisher = {Chapman and Hall/CRC},
+    year = {2020},
+    isbn = {9781138331457},
+    url = {https://plotly-r.com},
+  }
+
+@Manual{crosstalk,
+  title = {{crosstalk}: Inter-Widget Interactivity for HTML Widgets},
+  author = {Joe Cheng and Carson Sievert},
+  year = {2021},
+  note = {R package version 1.1.1},
+  url = {https://CRAN.R-project.org/package=crosstalk},
+}
+
+@article{RJ-2021-050,
+  author = {Earo Wang and Dianne Cook},
+  title = {Conversations in time: interactive visualisation to explore
+          structured temporal data},
+  year = {2021},
+  journal = {The R Journal},
+  doi = {10.32614/RJ-2021-050},
+  url = {https://journal.r-project.org/archive/2021/RJ-2021-050/index.html}
+}
+
+@Manual{palmerpenguins,
+  title = {{palmerpenguins}: Palmer Archipelago (Antarctica) penguin data},
+  author = {Allison Marie Horst and Alison Presmanes Hill and Kristen B Gorman},
+  year = {2020},
+  note = {R package version 0.1.0},
+  url = {https://allisonhorst.github.io/palmerpenguins/},
+}
+
 
 @misc{22arro,
 	title = {22  Arrow {\textendash} R for Data Science (2e)},
@@ -67,7 +102,6 @@ PMCID: PMC6331976},
 	author = {Dolladille, Charles and {Chrétien}, Basile},
 	year = {2025},
 	date = {2025},
-	doi = {10.32614/cran.package.vigicaen},
 	url = {https://github.com/pharmacologie-caen/vigicaen},
 	langid = {en}
 }
@@ -87,6 +121,96 @@ PMCID: PMC6331976},
 	year = {2023},
 	month = {02},
 	date = {2023-02-22},
-	doi = {10.32614/cran.package.tidyverse},
 	url = {https://cran.r-project.org/web/packages/tidyverse/index.html}
+}
+
+@article{legallois2025,
+	title = {Identification of anticancer drugs associated to cancer therapy-related cardiac dysfunction: a VigiBase® disproportionality analysis},
+	author = {Legallois, Damien and Da Silva, {Angélique} and Alexandre, Joachim and Milliez, Paul and Sabatier, {Rémi} and Blanchart, Katrien and Plane, Anne-Flore and Font, Jonaz and {Chrétien}, Basile and Dolladille, Charles},
+	year = {2025},
+	month = {08},
+	date = {2025-08-12},
+	journal = {European Heart Journal. Cardiovascular Pharmacotherapy},
+	pages = {459--468},
+	volume = {11},
+	number = {5},
+	doi = {10.1093/ehjcvp/pvaf027},
+	note = {PMID: 40272201
+PMCID: PMC12342995},
+	langid = {eng}
+}
+
+@article{dolladille2020,
+	title = {Immune Checkpoint Inhibitor Rechallenge After Immune-Related Adverse Events in Patients With Cancer},
+	author = {Dolladille, Charles and Ederhy, {Stéphane} and Sassier, Marion and Cautela, Jennifer and Thuny, Franck and Cohen, Ariel A. and Fedrizzi, Sophie and {Chrétien}, Basile and Da-Silva, {Angélique} and Plane, Anne-Flore and Legallois, Damien and Milliez, Paul U. and Lelong-Boulouard, {Véronique} and Alexandre, Joachim},
[…]
