### 2026-02-06 dafd604
bundle for 1.0.0
lifecycle, news update, version update, revdepcheck and cran comments update
