+++ revdep/library/vigicaen/new/purrr/doc/base.R
@@ -8,8 +8,6 @@ knitr::opts_chunk$set(
 )
 options(tibble.print_min = 6, tibble.print_max = 6)
 
-modern_r <- getRversion() >= "4.1.0"
-
 ## ----setup--------------------------------------------------------------------
 library(purrr)
 library(tibble)
@@ -45,7 +43,7 @@ str(samples)
 ## -----------------------------------------------------------------------------
 samples <- Map(function(...) rnorm(..., n = 5), mean = means, sd = sds)
 
-## ----eval = modern_r----------------------------------------------------------
+## -----------------------------------------------------------------------------
 samples <- Map(\(...) rnorm(..., n = 5), mean = means, sd = sds)
 
 ## -----------------------------------------------------------------------------
@@ -98,7 +96,7 @@ means |>
   map(rnorm, n = 5, sd = 1) |>
   map_dbl(median)
 
-## ----eval = modern_r----------------------------------------------------------
+## -----------------------------------------------------------------------------
 mtcars |>
   split(mtcars$cyl) |> 
   map(\(df) lm(mpg ~ wt, data = df))|> 
