### 2026-02-06 b57a67e
Deploying to gh-pages from @ pharmacologie-caen/vigicaen@dafd60493ef379b38f99c440b97340ad56628d49 🚀
