+++ reference/screen_drug.md
@@ -1,7 +1,7 @@
 # Screening of drugs
 
-**\[experimental\]** The `screen_drug()` function identifies and ranks
-the most frequently reported drugs (by active ingredient) in a dataset.
+**\[stable\]** The `screen_drug()` function identifies and ranks the
+most frequently reported drugs (by active ingredient) in a dataset.
 
 ## Usage
 
