+++ NEWS.md
@@ -1,3 +1,5 @@
+# vigicaen (development version)
+
 # vigicaen 1.0.0
 
 ## Time for 1.0!
