### 2026-02-18 69d9a64
Deploying to gh-pages from @ pharmacologie-caen/vigicaen@dafd60493ef379b38f99c440b97340ad56628d49 🚀
