### 2026-02-18 aa43244
Deploying to gh-pages from @ pharmacologie-caen/vigicaen@36cdc5fd5de8ff84c79d962d62531d5d0dd6b08a 🚀
