+++ tests/testthat/_snaps/tb_subset.md
@@ -33,7 +33,6 @@
       tb_subset(wd_in = paste0(wd_in, "/"), wd_out = paste0(wd_in, "/",
         "subset_drecno_item", "/"), subset_var = "drecno", sv_selection = sv_selection_drecno_item)
     Message
-      
       -- tb_subset() -----------------------------------------------------------------
       i Subsetting VigiBase tables.
       ====>-------------------------- percent, seconds | Reading source tables 
@@ -62,7 +61,6 @@
       tb_subset(wd_in = paste0(wd_in, "/"), wd_out = paste0(wd_in, "/", "subset_age",
         "/"), subset_var = "age", sv_selection = c(7, 8))
     Message
-      
       -- tb_subset() -----------------------------------------------------------------
       i Subsetting VigiBase tables.
       ====>-------------------------- percent, seconds | Reading source tables 
@@ -91,7 +89,6 @@
       tb_subset(wd_in = paste0(wd_in, "/"), wd_out = paste0(wd_in, "/",
         "subset_meddraid", "/"), subset_var = "meddra_id", sv_selection = sv_selection_mid)
     Message
-      
       -- tb_subset() -----------------------------------------------------------------
       i Subsetting VigiBase tables.
       ====>-------------------------- percent, seconds | Reading source tables 
@@ -121,7 +118,6 @@
         "subset_age_suspdup", "/"), subset_var = "age", sv_selection = c(7, 8),
       rm_suspdup = FALSE)
     Message
-      
       -- tb_subset() -----------------------------------------------------------------
       i Subsetting VigiBase tables.
       ====>-------------------------- percent, seconds | Reading source tables 
@@ -151,7 +147,6 @@
       tb_subset(wd_in = paste0(wd_in, "/"), wd_out = paste0(wd_in, "/", "subset_age",
         "/"), subset_var = "age", sv_selection = c(7, 8), rm_suspdup = TRUE)
     Message
-      
       -- tb_subset() -----------------------------------------------------------------
       i Subsetting VigiBase tables.
       ====>-------------------------- percent, seconds | Reading source tables 
@@ -180,7 +175,6 @@
       tb_subset(wd_in = wd_in, wd_out = paste0(wd_in, "/", "subset_age"), subset_var = "age",
       sv_selection = c(7, 8))
     Message
-      
       -- tb_subset() -----------------------------------------------------------------
       i Subsetting VigiBase tables.
       ====>-------------------------- percent, seconds | Reading source tables 
