### 2026-02-18 9702a0b
Deploying to gh-pages from @ pharmacologie-caen/vigicaen@fa9cbacd57e9d371996c6d835902ca82d6b7e0ad 🚀
