### 2026-03-17 #8935 @csoneson
👋🏻 @mhovd, @nanhung, @flor14 - would you be interested in reviewing this submission for the Journal of Open Source Software (JOSS)?

> Vigicaen: A vigibase® Pharmacovigilance Database Toolbox
https://github.com/openjournals/joss-reviews/issues/8935

Our checklist-driven reviews take place on GitHub, and review guidelines can be found [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

Many thanks in advance!
