### 2026-03-23 #8935 @csoneson
@mhovd, @nanhung, @flor14 - would you be available to help us review this submission for JOSS? Your expertise seem to fit very well. Thanks a lot in advance, and let me know if you have any questions.

> 👋🏻 [@mhovd](https://github.com/mhovd), [@nanhung](https://github.com/nanhung), [@flor14](https://github.com/flor14) - would you be interested in reviewing this submission for the Journal of Open Source Software (JOSS)?
>
> > Vigicaen: A vigibase® Pharmacovigilance Database Toolbox
> > [#8935](https://github.com/openjournals/joss-reviews/issues/8935)
>
> Our checklist-driven reviews take place on GitHub, and review guidelines can be found [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
>
> Many thanks in advance!
