### 2026-03-23 #8935 @mhovd
> [@mhovd](https://github.com/mhovd), [@nanhung](https://github.com/nanhung), [@flor14](https://github.com/flor14) - would you be available to help us review this submission for JOSS? Your expertise seem to fit very well. Thanks a lot in advance, and let me know if you have any questions.

Hi, I would be available and willing to review this submission :)
