### 2026-03-23 #8935 @flor14
Hello! I won't be available to review this package. Thank you so much for
the invitation!

El lun, 23 mar 2026 a las 9:32, Nan-Hung Hsieh ***@***.***>)
escribió:

> *nanhung* left a comment (openjournals/joss-reviews#8935)
> <https://github.com/openjournals/joss-reviews/issues/8935#issuecomment-4111980515>
>
> Hello, I am not familiar with the Pharmacovigilance Database and might not
> be suitable to review the package. Hoping you can find the ideal reviewer.
>
