### 2026-03-25 a4653d6
switch to csv input and rename MedicinalProd_Id to Record_Id
Aligns the package with the new VigiBase ECL export format.

- Updates tb_vigibase() and tb_who() to process .csv files instead of .txt.
- Replaces MedicinalProd_Id with Record_Id across functions, documentation, and tests.
- Removes the obsolete prg table and Pharmform_Id variable.
- Fixes a dplyr compatibility warning in get_llt_smq().
