+++ R/tb_vigibase.R
@@ -3,7 +3,7 @@
 #' @description `r lifecycle::badge('stable')` Transform VigiBase .txt
 #' files to .parquet files.
 #'
-#' @details Vigibase Extract Case Level is delivered as zipped text files, that you should
+#' @details Vigibase Extract Case Level is delivered as zipped csv files, that you should
 #' transform to a more efficient format. Parquet format from `arrow` has many advantages:
 #' It works with out-of-memory data, which makes it possible to process Vigibase tables on
 #' a computer with not-so-much RAM. It is also lightweighted and standard across different
@@ -17,7 +17,7 @@
 #' If set to FALSE (the default), the function will skip the construction of any .parquet tables that already exist,
 #' so you do not have to start from scratch after a failure. Set to TRUE to force rebuilding all tables.
 #'
-#' @param path_base Character string, a directory containing vigibase txt tables. It is also the
+#' @param path_base Character string, a directory containing vigibase csv tables. It is also the
 #' output directory.
 #' @param path_sub  Character string, a directory containing subsidiary tables.
 #' @param force Logical, to be passed to `cli::cli_progress_update()`. Used for internal
@@ -49,7 +49,7 @@
 #' @seealso [tb_who()], [tb_meddra()], [tb_subset()], [dt_parquet()]
 #'
 #' @return .parquet files into the `path_base` directory (**including suspected duplicates tables**).
-#' Some columns are returned as `integer` (UMCReportId, Drug_Id, MedicinalProd_Id, Adr_Id, MedDRA_Id),
+#' Some columns are returned as `integer` (UMCReportId, Drug_Id, Record_Id, Adr_Id, MedDRA_Id),
 #' and some columns as `numeric` (TimeToOnsetMin, TimeToOnsetMax)
 #' All other columns are `character`.
 #'
@@ -64,8 +64,8 @@
 #' dir.create(path_base)
 #' dir.create(path_sub)
 #'
-#' create_ex_main_txt(path_base)
-#' create_ex_sub_txt(path_sub)
+#' create_ex_main_csv(path_base)
+#' create_ex_sub_csv(path_sub)
 #'
 #'  # ---- Running tb_vigibase
 #'
@@ -93,17 +93,53 @@ tb_vigibase <-
       "tb_vigibase()"
     )
 
+    screen_csv_main <- tb_screen_main(path_base, ext = ".csv")
+    screen_csv_sub  <- tb_screen_sub( path_sub, ext = ".csv")
+
+    screen_csv <- c(screen_csv_main, screen_csv_sub)
+
+    expected_csv_files <-
+      paste0(
+        c("ADR", "DEMO", "DRUG", "FOLLOWUP", "IND", "LINK", "OUT", "SRCE",
+          "AgeGroup_Lx", "Dechallenge_Lx", "Dechallenge2_Lx", "Frequency_Lx",
+          "Gender_Lx", "Notifier_Lx", "Outcome_Lx", "Rechallenge_Lx", "Rechallenge2_Lx",
+          "Region_Lx", "ReportType_Lx", "RouteOfAdm_Lx", "Seriousness_Lx",
+          "SizeUnit_Lx", "SUSPECTEDDUPLICATES"),
+             ".csv")
+
+    if(all(expected_csv_files %in% screen_csv)) {
+      cli::cli_alert_success("All expected csv files found in {.arg path_base} and {.arg path_sub}")
+    } else {
+      missing_csv_files <-
+        expected_csv_files[!expected_csv_files %in% screen_csv]
+
+      if(length(missing_csv_files) == 8) { # 8 = number of tables in main
+        cli::cli_abort(
+          c(
+            "All csv files must be present in {.arg path_base} and {.arg path_sub}.",
[…]
