+++ tests/testthat/test-create_example_tables.R
@@ -4,52 +4,105 @@ test_that("written data is unchanged when loaded back", {
 
   dir.create(path)
 
-  create_ex_main_txt(path)
+  # from csv ----
 
-  main_reimport <-
-    names(f_sets_main()) |>
-    rlang::set_names() |>
+  create_ex_main_csv(path)
+
+  name_main_upper <- names(f_sets_main()) |>
+    stringr::str_to_upper()
+
+  name_main_csv <- paste0(name_main_upper, ".csv")
+
+  name_main_csv <-
+    ifelse(
+      name_main_csv == "SUSPDUP.csv",
+      "SUSPECTEDDUPLICATES.csv",
+      name_main_csv
+    )
+
+  main_reimport_csv <-
+    name_main_csv |>
+    rlang::set_names(names(f_sets_main())) |>
     purrr::map(function(tab_)
-      reader(tab_, path) |> dplyr::collect()
+      reader(tab_, path) |> dplyr::collect() |>
+        dplyr::mutate(dplyr::across(dplyr::everything(), as.character))
     )
 
   # beware that SRCE is exported both from main and who
 
   expect_equal(
-    main_reimport,
-    purrr::map(f_sets_main(), dplyr::as_tibble)
+    main_reimport_csv,
+    purrr::map(f_sets_main(), function(t_){
+      n_col_names <- ncol(t_) - 1
+      t_name <- names(t_) |>
+        rlang::set_names(paste0("f", 0:n_col_names))
+      t_ |> dplyr::as_tibble() |>
+        dplyr::mutate(dplyr::across(dplyr::everything(), as.character)) |>
+        # csv reading looses numeric/character type in case of numeric input
+        dplyr::rename(dplyr::all_of(t_name))
+    }
+    )
   )
 
   # sub
 
-  create_ex_sub_txt(path)
+  create_ex_sub_csv(path)
 
-  sub_reimport <-
-    names(f_sets_sub()) |>
-    rlang::set_names() |>
+  name_sub_upper <- names(f_sets_sub())
+  name_sub_csv <- paste0(name_sub_upper, ".csv")
+
+  sub_reimport_csv <-
+    name_sub_csv |>
+    rlang::set_names(names(f_sets_sub())) |>
     purrr::map(function(tab_)
-      reader(tab_, path) |> dplyr::collect()
+      reader(tab_, path) |> dplyr::collect() |>
+        dplyr::mutate(dplyr::across(dplyr::everything(), as.character))
     )
 
   expect_equal(
-    sub_reimport,
-    purrr::map(f_sets_sub(), dplyr::as_tibble)
+    sub_reimport_csv,
+    purrr::map(f_sets_sub(), function(t_){
+      n_col_names <- ncol(t_) - 1
+      t_name <- names(t_) |>
+        rlang::set_names(paste0("f", 0:n_col_names))
+      t_ |> dplyr::as_tibble() |>
+        dplyr::mutate(dplyr::across(dplyr::everything(), as.character)) |>
+        # csv reading looses numeric/character type in case of numeric input
+        dplyr::rename(dplyr::all_of(t_name))
+    }
+    )
   )
 
   # who
 
-  create_ex_who_txt(path)
+  create_ex_who_csv(path)
 
-  who_reimport <-
-    names(f_sets_who()) |>
-    rlang::set_names() |>
+  name_who_upper <- names(f_sets_who()) |>
+    stringr::str_to_upper()
+  name_who_csv <- paste0(name_who_upper, ".csv")
+
+  name_who_csv <- ifelse(name_who_csv == "UNITX.csv", "UNIT-X.csv", name_who_csv)
+
+  who_reimport_csv <-
+    name_who_csv |>
+    rlang::set_names(names(f_sets_who())) |>
     purrr::map(function(tab_)
[…]
