### 2026-03-25 7508fe1
hit 100% code cov - removed impossible if case
