### 2026-03-25 43145bf
correcting ThG naming for case sensitive systems
Windows does not seem to be case sensitive when reading files... But Linux is.
