### 2026-03-25 1713546
correcting case sensitive ThG and Unit-X
