### 2026-03-26 b7d0933
Deploying to gh-pages from @ pharmacologie-caen/vigicaen@97a72724d069fc5ed7da838c8a3f9c56d9426b9f 🚀
