+++ reference/get_atc_code.md
@@ -1,7 +1,7 @@
 # Get ATC codes (DrecNos or MPIs)
 
-**\[stable\]** Collect Drug Record Numbers or MedicinalProd_Ids
-associated to one or more ATC classes.
+**\[stable\]** Collect Drug Record Numbers or Record_Ids associated to
+one or more ATC classes.
 
 ## Usage
 
@@ -22,8 +22,8 @@ get_atc_code(atc_sel, mp, thg_data, vigilyze = TRUE)
 
 - thg_data:
 
-  A data.table. Correspondence between ATC codes and MedicinalProd_Id
-  (usually, it is `thg`)
+  A data.table. Correspondence between ATC codes and Record_Id (usually,
+  it is `thg`)
 
 - vigilyze:
 
@@ -33,18 +33,18 @@ get_atc_code(atc_sel, mp, thg_data, vigilyze = TRUE)
 ## Value
 
 A named list of integers. **DrecNos** if `vigilyze` is set to `TRUE`, or
-**MedicinalProd_Ids** if `vigilyze` is set to `FALSE`.
+**Record_Ids** if `vigilyze` is set to `FALSE`.
 
 ## Details
 
 `get_atc_code()` is an *ID collector* function. Provide `atc_sel` in the
 same way as `d_sel` in
 [`add_drug()`](https://pharmacologie-caen.github.io/vigicaen/reference/add_drug.md),
-but remember to specify its method arg as `MedicinalProd_Id` if
-`vigilyze` is set to `FALSE`. Vigilyze style means all conditioning of
-drugs will be retrieved after requesting an ATC class (i.e., drugs are
-identified with their DrecNos), even if a specific conditioning is not
-present in the ATC class. This is the default behavior in vigilyze.
+but remember to specify its method arg as `Record_Id` if `vigilyze` is
+set to `FALSE`. Vigilyze style means all conditioning of drugs will be
+retrieved after requesting an ATC class (i.e., drugs are identified with
+their DrecNos), even if a specific conditioning is not present in the
+ATC class. This is the default behavior in vigilyze.
 
 ## See also
 
@@ -74,12 +74,12 @@ atc_drecno <-
                vigilyze = TRUE)
 #> ℹ vigilyze set to TRUE, extracting DrecNos (?get_atc_code for details)
 
-# Or you can get MedicinalProd_Ids (if vigilyze is FALSE)
+# Or you can get Record_Ids (if vigilyze is FALSE)
 
 atc_mpi <-
   get_atc_code(atc_sel = atc_sel,
                mp = mp_,
                thg_data = thg_,
                vigilyze = FALSE)
-#> ℹ vigilyze set to FALSE, extracting MedicinalProd_ids (?get_atc_code for details)
+#> ℹ vigilyze set to FALSE, extracting Record_Ids (?get_atc_code for details)
 ```
