### 2026-03-28 #8935 @csoneson
@PatrickRWright, @pacoramon - would one of you be interested in reviewing this submission for the Journal of Open Source Software (JOSS)?

> Vigicaen: A vigibase® Pharmacovigilance Database Toolbox
https://github.com/openjournals/joss-reviews/issues/8935

Our checklist-driven reviews take place on GitHub, and review guidelines can be found [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

Many thanks in advance!
