### 2026-03-30 #10321 @editorialbot
[REVIEW]: 'Vigicaen: A vigibase® Pharmacovigilance Database Toolbox.
**Submitting author:** @cdolladille (<a href="http://orcid.org/0000-0003-0449-6261">Charles Dolladille</a>)
**Repository:** https://github.com/pharmacologie-caen/vigicaen
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** v2.1.0
