### 2026-03-30 #10321 @mhovd
> **Paper file info**:
>
> 🚨 Wordcount for `paper.md` is **2639**
> 🔴 Failed to discover a `State of the field` section in paper
> 🔴 Failed to discover a `Software design` section in paper
> 🔴 Failed to discover an `AI usage disclosure` section in paper

@csoneson - is the paper reviewed after the old or new criteria? In other words, are these sections required for this paper?
