### 2026-03-30 #10321 @csoneson
> > **Paper file info**:
> > 🚨 Wordcount for `paper.md` is **2639**
> > 🔴 Failed to discover a `State of the field` section in paper
> > 🔴 Failed to discover a `Software design` section in paper
> > 🔴 Failed to discover an `AI usage disclosure` section in paper
>
> [@csoneson](https://github.com/csoneson) - is the paper reviewed after the old or new criteria? In other words, are these sections required for this paper?

The paper is reviewed under the new scope criteria, and thus these sections should be included. However, I beleve the author was already asked to (and did) include them in the pre-review issue. Same with the missing DOIs.
@cdolladille - could you confirm that these aspects were addressed? Could it be that the updated paper is not anymore in the `joss_paper` branch (as indicated in the submission)?
