### 2026-03-31 #10321 @cdolladille
Thanks to all of you for your time in reviewing this manuscript. 🙏🏼
I did add new sections as requested during the pre review process, including Research impact and significance, Design Thinking, Open Source Software practice, development history.
Possibly due to an unexpected omission, the State of the field, Software design and AI usage disclosure sections were not added in the meantime.

I will add these sections as quickly as I can and I sincerely apologize for any inconvenience this night have caused.
