### 2026-04-11 #10321 @csoneson
Hi @cdolladille - just wanted to quickly check in regarding the updates of the paper with the new sections (to give the reviewers an up-to-date starting point). Let us know if anything is unclear. Thanks!
