### 2026-04-23 #10321 @csoneson
Hi @cdolladille - just wanted to check in to see if you have an ETA of the updates to the paper. Thanks.
