### 2026-06-17 65f3455
Add contributing guidelines (#169)
Introduce a CONTRIBUTING.md file to provide clear instructions for
how to propose changes, report bugs, and adhere to project
conventions for both new and existing contributors.
