### 2026-06-17 5c36da6
Fixing References mistakes and imprecisions (#171)
