### 2026-06-17 c4e9b83
Fixing References mistakes and imprecisions (#171)
