+++ paper.bib
@@ -34,35 +34,49 @@
   url = {https://allisonhorst.github.io/palmerpenguins/},
 }
 
-
-@misc{22arro,
-	title = {22  Arrow {\textendash} R for Data Science (2e)},
-	url = {https://r4ds.hadley.nz/arrow.html}
+@online{22arro,
+  author = {Wickham, Hadley and Çetinkaya-Rundel, Mine and Grolemund, Garrett},
+  title = {Arrow -- R for Data Science},
+  year = {2023},
+  edition = {2e},
+  url = {https://r4ds.hadley.nz/arrow.html},
+  urldate = {2026-06-17}
 }
 
-@misc{parquet,
-	title = {Parquet},
-	url = {https://parquet.apache.org/},
-	langid = {en}
+@online{parquet,
+  author = {{Apache Software Foundation}},
+  title = {Apache Parquet},
+  year = {2026},
+  url = {https://parquet.apache.org/},
+  urldate = {2026-06-17},
+  langid = {en}
 }
 
-@misc{apachea,
-	title = {Apache Arrow},
-	url = {https://arrow.apache.org/},
-	langid = {canadian}
+@online{apachea,
+  author = {{Apache Software Foundation}},
+  title = {Apache Arrow},
+  year = {2026},
+  url = {https://arrow.apache.org/},
+  urldate = {2026-06-17},
+  langid = {canadian}
 }
 
-@misc{arrowr,
-	title = {Arrow R Package},
-	url = {https://arrow.apache.org/docs/r/},
-	langid = {canadian}
+@Manual{arrowr,
+    title = {arrow: Integration to 'Apache' 'Arrow'},
+    author = {Neal Richardson and Ian Cook and Nic Crane and Dewey Dunnington and Romain François and Jonathan Keane and Bryce Mecum and Dragoș Moldovan-Grünfeld and Jeroen Ooms and Jacob Wujciak-Jens and {Apache Arrow}},
+    year = {2026},
+    note = {R package version 23.0.1.2},
+    url = {https://CRAN.R-project.org/package=arrow},
+    doi = {10.32614/CRAN.package.arrow},
 }
 
-@misc{vigibase,
-	title = {About VigiBase},
-	author = {Centre, Uppsala Monitoring},
-	url = {https://who-umc.org/vigibase/},
-	langid = {en}
+@online{vigibase,
+  author = {{Uppsala Monitoring Centre}},
+  title = {About VigiBase},
+  year = {2026},
+  url = {https://who-umc.org/vigibase/},
+  urldate = {2026-06-17},
+  langid = {en}
 }
 
 @article{montastruc2011,
@@ -210,7 +224,55 @@ PMID: 40617160},
 	langid = {eng}
 }
 
-@misc{rpackag,
-	title = {R Packages (2e)},
-	url = {https://r-pkgs.org/}
+@book{rpackag,
+  author    = {Wickham, Hadley and Bryan, Jenny},
+  title     = {R Packages: Organize, Test, Document, and Share Your Code},
+  year      = {2023},
+  edition   = {2e},
+  publisher = {O'Reilly Media},
+  url       = {https://r-pkgs.org/},
+  urldate   = {2026-06-17},
+  isbn      = {978-1098134464},
+  langid    = {en}
 }
+
+@article{mukhopadhyay2026,
+	title = {{SurVigilance}: {An} application for accessing global pharmacovigilance data},
+	volume = {34},
+	issn = {2352-7110},
+	shorttitle = {{SurVigilance}},
+	url = {https://www.sciencedirect.com/science/article/pii/S2352711026000403},
+	doi = {10.1016/j.softx.2026.102546},
[…]
