### 2026-06-17 00f6384
wip fixing typos and spelling errors
Query in this post: https://github.com/openjournals/joss-reviews/issues/10321#issuecomment-4155080703
