+++ paper.md
@@ -49,16 +49,16 @@ analyses using pharmacovigilance data, as traditional approaches are
 susceptible to various biases such as reporting bias and confounding.
 The aim of vigicaen is to provide a toolbox for the VigiBase® Extract
 Case Level database, resolving technical challenges related to the
-database large size, and providing easier and reproducible access to
+large size of the database, and providing easier and reproducible access to
 advanced features. The package is built on top of the parquet file
-format. Functions related to drug and adverse event identification,
-descriptive features such as time to onset, dechallenge and rechallenge
-outcomes are provided. Command line side-effect outputs aim at fast
-resolving of common issues related to drug and adverse event
+format. Functions related to drug and adverse event identification, and
+descriptive features such as time to onset, dechallenge, and rechallenge
+outcomes are provided. Command-line side-effect outputs aim at fast
+resolution of common issues related to drug and adverse event
 identification. The package is intended for pharmacovigilance
-practitioners, clinicians and researchers with or without advanced
+practitioners, clinicians, and researchers with or without advanced
 biostatistical skills. A graphical output can be produced for routine
-use, to support daily assessment of drug liability.
+use to support the daily assessment of drug liability.
 
 # Statement of need
 
@@ -72,12 +72,12 @@ of pharmacovigilance principles and reporting systems.
 
 For decades, the World Health Organization (WHO) has been collecting
 adverse drug reaction reports, called Individual Case Safety Reports
-(ICSRs), from its member countries, populating more than 40 millions
+(ICSRs), from its member countries, populating more than 40 million
 reports to date. This pharmacovigilance database is called VigiBase® and
 is managed by the Uppsala Monitoring Centre in Sweden.[@vigibase] These
 ICSRs describe the course of patients who experienced an adverse event
 (a medical condition) after taking a drug. The burning question is
-whether this adverse event was actually related to the drug intake, e.g.
+whether this adverse event was actually related to the drug intake, e.g.,
 if it is an adverse drug *reaction* (ADR). The pharmacovigilance
 database aims at uncovering the very first potential signals of
 association between drugs and ADRs.[@montastruc2011]
@@ -85,38 +85,38 @@ association between drugs and ADRs.[@montastruc2011]
 It relies on disproportionality analysis, a statistical method that
 produces estimators of how unlikely the number of observed ICSRs
 reporting on a specific drug and adverse event is to be attributable to
-chance alone. Together with an incertitude margin, these estimators are
+chance alone. Together with an uncertainty margin, these estimators are
 used to raise safety signals on drugs.[@montastruc2011]
 
 The Uppsala Monitoring Centre grants access to VigiBase® to researchers,
-either academics or industrials, under a licence contract. The most
-extensive available version is called Extract Case Level: It contains
-all the ICSRs, with information such as the patient demographics, the
-drug intake, the adverse events, the outcome, the dechallenge and
-rechallenge outcome, and the time to onset. However, this version is
-provided as large text files, and requires a lot of processing before
-being usable for analysis. Those text files might be particularly
-challenging to use in R, as they would often exceed the size of the
-available Random Access Memory, thus requiring advanced knowledge of R
-computing techniques. Clinicians and pharmacovigilance practitioners
-typically lack these skills, and therefore struggle to use the VigiBase®
-data for their research. As a result, they would often rely on partial
-data, with limited statistical modelling options. Or, they could develop
-home-made biostatistic scripts that would typically be used once, often
-left undocumented, and highly heterogeneous across research teams.
+either academic or industrial, under a license contract. The most
+extensive available version is called Extract Case Level: it contains
+all the ICSRs, with information such as patient demographics, drug intake,
[…]
