### 2026-06-23 44086ed
Removing blank lines in code chunks in latex version
