### 2026-06-23 7e6fcbf
Adding details on disproportionality estimates purpose
And improved consistency of reference style
