### 2026-06-24 32d1322
Add duplicate names check to check_id_list with tests and NEWS entry (#142)
