### 2026-06-24 d92ace5
Rename dn to dup_name for clarity in check_id_list
