### 2026-06-24 00e33f1
Fix get_* functions failing with Arrow Dataset (open_dataset) tables (#167)
