### 2026-06-24 c6470ad
feat: configure roxygen2 version and remove RoxygenNote
