+++ DESCRIPTION
@@ -25,7 +25,6 @@ Encoding: UTF-8
 LazyData: true
 LazyDataCompression: xz
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.3.3
 URL: https://github.com/pharmacologie-caen/vigicaen, https://pharmacologie-caen.github.io/vigicaen/
 BugReports: https://github.com/pharmacologie-caen/vigicaen/issues
 Suggests: 
@@ -50,3 +49,4 @@ Imports:
     stringr,
     tidyr
 VignetteBuilder: knitr
+Config/roxygen2/version: 8.0.0
