### 2026-06-24 6577bda
test: refactor add_dose test with complex input and update snapshots
