### 2026-06-24 8378325
test: enhance check_id_list with duplicate name and CLI format tests
