### 2026-06-24 554799c
test: verify add_dose can overwrite existing columns
hit back 100% coverage
