### 2026-06-24 96c6730
Deploying to gh-pages from @ pharmacologie-caen/vigicaen@5320abd38f5c4d645d95c3408806de9d7f07e1f8 🚀
