### 2026-06-24 a453b54
Deploying to gh-pages from @ pharmacologie-caen/vigicaen@e5a5afadec567b0c7f40a967c8231d34f97b0fd5 🚀
