### 2026-06-24 5aacdcf
Increment version number to 2.0.0
