### 2026-06-24 905102d
Deploying to gh-pages from @ pharmacologie-caen/vigicaen@5bba27c75bb652efbadac69cd9ca2000ef6177a8 🚀
