+++ revdep/library/vigicaen/old/vigicaen/doc/template_main.R
@@ -1,291 +1,0 @@
-## ----setup, include = FALSE---------------------------------------------------
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>",
-  include = TRUE,
-  echo = TRUE,
-  warning = FALSE, 
-  message = FALSE
-)
-
-## ----libraries----------------------------------------------------------------
-library(vigicaen)
-library(rlang)
-library(here) # if you like the here syntax for file paths
-library(dplyr)
-
-## ----path_classic, eval = FALSE-----------------------------------------------
-# # #### PATHS #### ####
-# 
-# path_base   <- "../vigibase_ecl/main"
-# 
-# # #### IMPORT #### ####
-# 
-# demo     <- dt_parquet(path_base, "demo", in_memory = FALSE)
-# adr      <- dt_parquet(path_base, "adr",  in_memory = FALSE)
-# drug     <- dt_parquet(path_base, "drug", in_memory = FALSE)
-# link     <- dt_parquet(path_base, "link", in_memory = FALSE)
-# out      <- dt_parquet(path_base, "out",  in_memory = FALSE)
-# srce     <- dt_parquet(path_base, "srce", in_memory = FALSE)
-# followup <- dt_parquet(path_base, "followup", in_memory = FALSE)
-# 
-# suspdup  <- dt_parquet(path_base, "suspdup", in_memory = FALSE)
-# 
-# source("00_dict.R") # from template_dictionary
-
-## ----test_datasets, include = FALSE-------------------------------------------
-demo     <- demo_
-adr      <- adr_
-drug     <- drug_
-link     <- link_
-out      <- out_
-followup <- followup_
-srce     <- srce_
-
-thg      <- thg_
-mp       <- mp_
-meddra   <- meddra_
-smq_list <- smq_list_
-smq_content <- smq_content_
-
-suspdup <- 
-  data.table::data.table(
-    UMCReportId = 1,
-    SuspectedduplicateReportId = NA
-  )
-
-## ----codes, include = FALSE, eval=TRUE, message = FALSE-----------------------
-d_drecno <- ex_$d_drecno
-
-atc_drecno <-
-  get_atc_code(atc_sel =  rlang::list2(l03 = c("L03AA")),
-               mp = mp_,
-               thg_data = thg_,
-               vigilyze = TRUE)
-
-a_llt <- ex_$a_llt
-
-## ----demo_dm, eval = TRUE-----------------------------------------------------
-# ---- Deduplicating ---- ####
-
-demo <- demo |> 
-  filter(!(UMCReportId %in% suspdup$SuspectedduplicateReportId))
-
-# ---- Drugs ---- ####
-
-# From a list of drugs
-
-demo <-
-  demo |>
-  add_drug(
-    d_code = d_drecno,
-    drug_data = drug
-  )
-
-# From ATC
-
-demo <-
-  demo |>
-  add_drug(
-    d_code = atc_drecno,
-    drug_data = drug
-  )
-
-# ---- Reactions ---- ####
-
-demo <-
-  demo |>
-  add_adr(
-    a_code = a_llt,
-    adr_data = adr
-  )
-
-# ---- Demographics ---- ####
-
-# Age, sex
-
-demo <-
-  demo |>
-  mutate(
-    age = cut(as.integer(AgeGroup),
-              breaks = c(0,4,5,6,7,8),
-              include.lowest = TRUE, right = TRUE,
-              labels = c("<18", "18-45","45-64", "65-74", "75+")),
-
-    sex = case_when(Gender == "1" ~ 1,
-                    Gender == "2" ~ 2,
-                    Gender %in% c("-","0","9") ~ NA_real_,
[…]
