+++ revdep/library/vigicaen/old/vigicaen/doc/basic_workflow.R
@@ -1,298 +1,0 @@
-## ----include = FALSE----------------------------------------------------------
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>"
-)
-
-## ----library, warning=FALSE, message=FALSE------------------------------------
-library(vigicaen)
-library(rlang)
-library(dplyr)
-
-## ----named_lists--------------------------------------------------------------
-# drug selection
-d_sel <-
-  list(
-    nivolumab   = "nivolumab",
-    ipilimumab  = "ipilimumab",
-    nivo_or_ipi = c("nivolumab", "ipilimumab")
-  )
-
-# adverse drug reaction selection
-
-a_sel <- 
-  list(
-    colitis     = "Colitis",
-    pneumonitis = "Pneumonitis"
-  )
-
-## ----table_loading, warning=FALSE, message=FALSE------------------------------
-demo <- demo_
-drug <- drug_
-
-mp <- mp_
-
-## ----drug_sel-----------------------------------------------------------------
-d_sel <- # drug selection
-  list2(
-    nivolumab = "nivolumab"
-  )
-
-d_sel
-
-## ----verbose_with_get_drecno--------------------------------------------------
-get_drecno(
-  d_sel = d_sel,
-  mp = mp_,
-  verbose = TRUE
-  )
-
-## ----d_drecno-----------------------------------------------------------------
-d_drecno <-
-  get_drecno(
-    d_sel = d_sel,
-    mp = mp_,
-    verbose = FALSE
-    )
-
-## ----add_drug-----------------------------------------------------------------
-demo <- 
-  add_drug(
-    .data = demo,
-    d_code = d_drecno,
-    drug_data = drug)
-demo
-
-## ----results='hide'-----------------------------------------------------------
-demo <- 
-  demo |> 
-  add_drug(
-    d_code = d_drecno,
-    drug_data = drug
-  )
-
-## ----check_dm-----------------------------------------------------------------
-check_dm(demo, "nivolumab")
-
-## ----atc_drecno---------------------------------------------------------------
-atc_sel <-
-  list2(l03 = "L03")
-
-atc_drecno <- 
-  get_atc_code(
-    atc_sel = atc_sel,
-    mp = mp,
-    thg_data = thg_
-    )
-
-## ----str_atc_drecno-----------------------------------------------------------
-str(atc_drecno)
-
-## ----add_atc------------------------------------------------------------------
-demo |> 
-  add_drug(
-    d_code = atc_drecno,
-    drug_data = drug
-  )
-
-## ----add_drug_repbasis--------------------------------------------------------
-demo |> 
-  add_drug(
-    d_code = d_drecno,
-    drug_data = drug,
-    repbasis = "sci"
-  ) |> 
-  check_dm("nivolumab")
-
-# suspected only
-
-demo |> 
-  add_drug(
-    d_code = d_drecno,
-    drug_data = drug,
-    repbasis = "s"
-  ) |> 
-  check_dm("nivolumab")
-
-## ----many_drugs---------------------------------------------------------------
-d_sel <- 
-  list2(
-    nivolumab = "nivolumab",
-    pembrolizumab = "pembrolizumab"
-  )
-
-d_drecno <-
-  d_sel |> 
-  get_drecno(mp = mp)
-
-demo <- 
-  demo |> 
-  add_drug(
-    d_drecno,
-    drug_data = drug
-  )
-
-demo |> 
-  check_dm(c("nivolumab", "pembrolizumab"))
-
-## ----d_groups-----------------------------------------------------------------
-d_sel <- 
-  list2(
-    analgesics = c("paracetamol", "tramadol"),
-    ici = c("nivolumab", "pembrolizumab")
-  )
-
-d_drecno <-
-  d_sel |> 
-  get_drecno(mp = mp,
[…]
