### 2026-06-24 337e60b
Deploying to gh-pages from @ pharmacologie-caen/vigicaen@ae05d9be5719f6c6c9391ca85986bc986b1fb3eb 🚀
