+++ NEWS.md
@@ -1,3 +1,5 @@
+# vigicaen (development version)
+
 # vigicaen 2.0.0
 
 ## Breaking changes
