+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: vigicaen
 Title: 'VigiBase' Pharmacovigilance Database Toolbox
-Version: 2.0.0
+Version: 2.0.0.9000
 Authors@R: c(
        person("Charles", "Dolladille", email = "cdolladille@hotmail.com", role = c("aut", "cre"),
            comment = c(ORCID = "0000-0003-0449-6261")),
