### 2026-06-25 8e1f408
Deploying to gh-pages from @ pharmacologie-caen/vigicaen@ae05d9be5719f6c6c9391ca85986bc986b1fb3eb 🚀
