+++ news/index.md
@@ -2,6 +2,8 @@
 
 ## vigicaen 2.0.0
 
+CRAN release: 2026-06-24
+
 ### Breaking changes
 
 - Due to a major update of the vigibase extract format (now delivered as
