+++ reference/vigi_routine.md
@@ -184,5 +184,5 @@ vigi_routine(
   export_to = paste0(tempdir(), "/", "vigicaen_graph.png")
 )
 
-#> Plot exported to /tmp/Rtmp5rlQvE/vigicaen_graph.png
+#> Plot exported to /tmp/RtmpR2fcoi/vigicaen_graph.png
 ```
