### 2026-06-25 d20dfd0
Deploying to gh-pages from @ pharmacologie-caen/vigicaen@dc2bccc3dcbc24e1a911505cf04fddda08a8410d 🚀
