### 2026-06-25 db041b9
feat: add tidy-select support to desc functions
