### 2026-06-25 5414c75
feat: add verbose success output to get_atc_code and get_llt_smq
