### 2026-06-25 f60e8ed
test: avoid duplicate verbose call in get_atc_code test
