### 2026-06-25 caa79b6
Add add_death, add_fup, add_serious functions for outcome data management (#144)
