+++ tests/testthat/test-add_outcomes.R
@@ -1,0 +1,546 @@
+test_that("add_death works with in-memory tables", {
+
+  out_test <- data.table::data.table(
+    UMCReportId  = c(1L, 2L, 3L),
+    Seriousness  = c("1", "6", "-"),
+    Serious      = c("Y", "Y", "N")
+  )
+
+  demo_test <- data.table::data.table(
+    UMCReportId  = c(1L, 2L, 3L, 4L),
+    Region       = NA_character_,
+    DateDatabase = NA_character_,
+    Type         = NA_character_
+  )
+
+  result <- add_death(demo_test, out_data = out_test)
+
+  expect_equal(ncol(result), ncol(demo_test) + 1L)
+  expect_true("death" %in% names(result))
+
+  # UMCReportId 1: in out, Seriousness == "1" → TRUE
+  expect_true(result[result$UMCReportId == 1L, ]$death)
+  # UMCReportId 2: in out, Seriousness != "1" → FALSE
+  expect_false(result[result$UMCReportId == 2L, ]$death)
+  # UMCReportId 3: in out, Seriousness == "-" → FALSE
+  expect_false(result[result$UMCReportId == 3L, ]$death)
+  # UMCReportId 4: not in out → NA
+  expect_true(is.na(result[result$UMCReportId == 4L, ]$death))
+})
+
+test_that("add_death works with Arrow tables", {
+
+  out_test <- data.table::data.table(
+    UMCReportId  = c(1L, 2L, 3L),
+    Seriousness  = c("1", "6", "-"),
+    Serious      = c("Y", "Y", "N")
+  )
+
+  demo_test <- data.table::data.table(
+    UMCReportId  = c(1L, 2L, 3L, 4L),
+    Region       = NA_character_,
+    DateDatabase = NA_character_,
+    Type         = NA_character_
+  )
+
+  result_arrow <-
+    add_death(
+      arrow::as_arrow_table(demo_test),
+      out_data = arrow::as_arrow_table(out_test)
+    ) |>
+    dplyr::collect()
+
+  expect_equal(ncol(result_arrow), ncol(demo_test) + 1L)
+  expect_true("death" %in% names(result_arrow))
+
+  expect_true(result_arrow[result_arrow$UMCReportId == 1L, ]$death)
+  expect_false(result_arrow[result_arrow$UMCReportId == 2L, ]$death)
+  expect_false(result_arrow[result_arrow$UMCReportId == 3L, ]$death)
+  expect_true(is.na(result_arrow[result_arrow$UMCReportId == 4L, ]$death))
+})
+
+test_that("add_death works with preloaded data", {
+
+  result <- add_death(demo_, out_data = out_)
+
+  expect_equal(ncol(result), ncol(demo_) + 1L)
+  expect_true("death" %in% names(result))
+  expect_true(sum(result$death, na.rm = TRUE) > 0)
+  expect_true(any(is.na(result$death)))
+})
+
+test_that("add_death col_name argument works", {
+
+  out_test <- data.table::data.table(
+    UMCReportId  = c(1L, 2L),
+    Seriousness  = c("1", "6"),
+    Serious      = c("Y", "Y")
+  )
+
+  demo_test <- data.table::data.table(
+    UMCReportId  = c(1L, 2L, 3L),
+    Region       = NA_character_,
+    DateDatabase = NA_character_,
+    Type         = NA_character_
+  )
+
+  result <- add_death(demo_test, out_data = out_test, col_name = "my_death")
+
+  expect_true("my_death" %in% names(result))
+  expect_false("death" %in% names(result))
+})
+
+test_that("add_death raises error on wrong out_data", {
+
[…]
