+++ NAMESPACE
@@ -1,9 +1,12 @@
 # Generated by roxygen2: do not edit by hand
 
 export(add_adr)
+export(add_death)
 export(add_dose)
 export(add_drug)
+export(add_fup)
 export(add_ind)
+export(add_serious)
 export(cff)
 export(check_dm)
 export(compute_dispro)

[…]
