### 2026-06-25 b8dfde0
Fix duplicate test blocks in test-add_outcomes.R
