### 2026-06-25 e4a02f1
Address PR review: query_data_type check, arrow.pull_as_vector option, 1/0 output, glue column naming, new tests
