+++ tests/testthat/test-add_outcomes.R
@@ -1,3 +1,45 @@
+test_that("check_data_out validates outcome data structure", {
+
+  out_valid <- data.frame(
+    UMCReportId = 1L, Seriousness = "1", Serious = "Y"
+  )
+
+  expect_null(
+    vigicaen:::check_data_out(out_valid, ".data")
+  )
+
+  expect_snapshot(
+    error = TRUE,
+    vigicaen:::check_data_out(data.frame(UMCReportId = 1L), ".data")
+  )
+
+  expect_snapshot(
+    error = TRUE,
+    vigicaen:::check_data_out(demo_, ".data")
+  )
+})
+
+test_that("check_data_fup validates followup data structure", {
+
+  fup_valid <- data.frame(
+    UMCReportId = 1L, ReplacedUMCReportId = 2L
+  )
+
+  expect_null(
+    vigicaen:::check_data_fup(fup_valid, ".data")
+  )
+
+  expect_snapshot(
+    error = TRUE,
+    vigicaen:::check_data_fup(data.frame(UMCReportId = 1L), ".data")
+  )
+
+  expect_snapshot(
+    error = TRUE,
+    vigicaen:::check_data_fup(out_, ".data")
+  )
+})
+
 test_that("add_death works with in-memory tables", {
 
   out_test <- data.table::data.table(
@@ -13,17 +55,17 @@ test_that("add_death works with in-memory tables", {
     Type         = NA_character_
   )
 
-  result <- add_death(demo_test, out_data = out_test)
+  result <- suppressMessages(add_death(demo_test, out_data = out_test))
 
   expect_equal(ncol(result), ncol(demo_test) + 1L)
   expect_true("death" %in% names(result))
 
-  # UMCReportId 1: in out, Seriousness == "1" → TRUE
-  expect_true(result[result$UMCReportId == 1L, ]$death)
-  # UMCReportId 2: in out, Seriousness != "1" → FALSE
-  expect_false(result[result$UMCReportId == 2L, ]$death)
-  # UMCReportId 3: in out, Seriousness == "-" → FALSE
-  expect_false(result[result$UMCReportId == 3L, ]$death)
+  # UMCReportId 1: in out, Seriousness == "1" → 1
+  expect_equal(result[result$UMCReportId == 1L, ]$death, 1L)
+  # UMCReportId 2: in out, Seriousness != "1" → 0
+  expect_equal(result[result$UMCReportId == 2L, ]$death, 0L)
+  # UMCReportId 3: in out, Seriousness == "-" → 0
+  expect_equal(result[result$UMCReportId == 3L, ]$death, 0L)
   # UMCReportId 4: not in out → NA
   expect_true(is.na(result[result$UMCReportId == 4L, ]$death))
 })
@@ -44,24 +86,26 @@ test_that("add_death works with Arrow tables", {
   )
 
   result_arrow <-
-    add_death(
-      arrow::as_arrow_table(demo_test),
-      out_data = arrow::as_arrow_table(out_test)
+    suppressMessages(
+      add_death(
+        arrow::as_arrow_table(demo_test),
+        out_data = arrow::as_arrow_table(out_test)
+      )
     ) |>
     dplyr::collect()
 
   expect_equal(ncol(result_arrow), ncol(demo_test) + 1L)
   expect_true("death" %in% names(result_arrow))
 
-  expect_true(result_arrow[result_arrow$UMCReportId == 1L, ]$death)
-  expect_false(result_arrow[result_arrow$UMCReportId == 2L, ]$death)
-  expect_false(result_arrow[result_arrow$UMCReportId == 3L, ]$death)
+  expect_equal(result_arrow[result_arrow$UMCReportId == 1L, ]$death, 1L)
[…]
