### 2026-06-25 e0efc4e
fixes to unit tests
