### 2026-06-25 9c2f70e
test: cover resolve_desc_vars behavior
