+++ tests/testthat/test-desc_cont.R
@@ -181,6 +181,48 @@ test_that("tidy-select works", {
   )
 })
 
+test_that("resolve_desc_vars preserves legacy inputs and tidy-select", {
+  df <-
+    data.frame(
+      smoke_status = c("smoker", "non-smoker"),
+      age = c(60, 50),
+      bmi = c(18, 30)
+    )
+
+  named_vc <- c(primary = "age")
+  list_vc <- list(primary = "age")
+  missing_vc <- c(primary = "weight")
+
+  expect_equal(
+    vigicaen:::resolve_desc_vars(df, rlang::quo(age), col_arg = "vc"),
+    "age"
+  )
+
+  expect_equal(
+    vigicaen:::resolve_desc_vars(
+      df,
+      rlang::quo(dplyr::starts_with("b")),
+      col_arg = "vc"
+    ),
+    "bmi"
+  )
+
+  expect_equal(
+    vigicaen:::resolve_desc_vars(df, rlang::quo(named_vc), col_arg = "vc"),
+    "age"
+  )
+
+  expect_equal(
+    vigicaen:::resolve_desc_vars(df, rlang::quo(list_vc), col_arg = "vc"),
+    "age"
+  )
+
+  expect_error(
+    vigicaen:::resolve_desc_vars(df, rlang::quo(missing_vc), col_arg = "vc"),
+    class = "columns_not_in_data"
+  )
+})
+
 test_that(
   "missing data", {
     df <-
