### 2026-06-25 63b6c66
fixes to r cmd check
