+++ NAMESPACE
@@ -51,6 +51,7 @@ importFrom(data.table,setDTthreads)
 importFrom(glue,glue_collapse)
 importFrom(gridExtra,grid.arrange)
 importFrom(lifecycle,deprecated)
+importFrom(rlang,":=")
 importFrom(rlang,.data)
 importFrom(rlang,.env)
 importFrom(stats,median)
