### 2026-06-25 b13500b
refactor: narrow list selector support
