+++ R/desc_tto.R
@@ -64,8 +64,8 @@ desc_tto <-
            ...
            ){
 
-    adr_s <- resolve_desc_vars(.data, rlang::enquo(adr_s), col_arg = "adr_s")
-    drug_s <- resolve_desc_vars(.data, rlang::enquo(drug_s), col_arg = "drug_s")
+    adr_s <- resolve_desc_vars(.data, rlang::enquo(adr_s), col_arg = "adr_s", allow_list = TRUE)
+    drug_s <- resolve_desc_vars(.data, rlang::enquo(drug_s), col_arg = "drug_s", allow_list = TRUE)
 
     # extract ttos
 
