### 2026-06-25 8a66432
test: tighten selector helper coverage
