### 2026-06-25 ce99252
fix: refine verbose sections for get_atc_code and get_llt_smq
