+++ tests/testthat/test-get_atc_code.R
@@ -10,15 +10,12 @@ test_that("Works for single and multiple ATC selections", {
                  c09aa = 26
     )
 
-  expect_message({
   atc_drecno <-
     get_atc_code(atc_sel = atc_sel,
                  mp = mp_,
                  thg_data = thg_,
-                 vigilyze = TRUE)
-  },
-  "DrecNo"
-  )
+                 vigilyze = TRUE,
+                 verbose = FALSE)
 
   purrr::iwalk(
     atc_sel_drecno_counts,
@@ -42,23 +39,20 @@ test_that("extracts record_id if vigilyze is FALSE", {
                  c09aa = 2185
     )
 
-  expect_message({
-    atc_drecno <-
-      get_atc_code(atc_sel = atc_sel,
-                   mp = mp_,
-                   thg_data = thg_,
-                   vigilyze = FALSE)
+  atc_drecno <-
+    get_atc_code(atc_sel = atc_sel,
+                 mp = mp_,
+                 thg_data = thg_,
+                 vigilyze = FALSE,
+                 verbose = FALSE)
 
-    purrr::iwalk(
-      atc_sel_mpi_counts,
-      function(a, a_n)
-        expect_equal(
-          length(atc_drecno[[a_n]]),
-          a
-        )
-    )
-  },
-  "Record_Id"
+  purrr::iwalk(
+    atc_sel_mpi_counts,
+    function(a, a_n)
+      expect_equal(
+        length(atc_drecno[[a_n]]),
+        a
+      )
   )
 
 })
@@ -76,17 +70,14 @@ test_that("names are tolower-ed and trimed", {
 
   # with vigilyze to TRUE
 
-  expect_warning({
-    expect_message({
+  expect_warning(
     atc_drecno <<-
       get_atc_code(atc_sel = atc_sel,
                    mp = mp_,
                    thg_data = thg_,
-                   vigilyze = TRUE)
-  },
-  "DrecNo"
-  )
-  }, "tolower-ed"
+                   vigilyze = TRUE,
+                   verbose = FALSE),
+    "tolower-ed"
   )
 
   expect_equal(
@@ -96,19 +87,17 @@ test_that("names are tolower-ed and trimed", {
 
   # with vigilyze to FALSE
 
-  expect_warning({
-    expect_message({
-      atc_drecno_vigifalse <<-
-        get_atc_code(
-          atc_sel = atc_sel,
-          mp = mp_,
-          thg_data = thg_,
-          vigilyze = FALSE
-        )
-    },
-    "Record_Id")
-  },
-  "tolower-ed")
+  expect_warning(
+    atc_drecno_vigifalse <<-
+      get_atc_code(
+        atc_sel = atc_sel,
+        mp = mp_,
+        thg_data = thg_,
+        vigilyze = FALSE,
+        verbose = FALSE
+      ),
+    "tolower-ed"
+  )
 
   expect_equal(
     names(atc_drecno_vigifalse),
@@ -128,17 +117,14 @@ test_that("works with thg or mp as Table (out of memory)", {
                  c09aa = 26
     )
 
-  expect_message({
-    atc_drecno <-
-      get_atc_code(atc_sel = atc_sel,
-                   mp = mp_ |>
-                     arrow::as_arrow_table(),
-                   thg_data = thg_ |>
-                     arrow::as_arrow_table(),
-                   vigilyze = TRUE)
-  },
-  "DrecNo"
-  )
+  atc_drecno <-
+    get_atc_code(atc_sel = atc_sel,
+                 mp = mp_ |>
+                   arrow::as_arrow_table(),
[…]
