### 2026-06-25 5b39198
Remove list selector support
