+++ R/desc_cont.R
@@ -277,8 +277,7 @@ resolve_desc_vars <-
     function(.data,
              cols_quo,
              col_arg = rlang::as_label(cols_quo),
-             call = rlang::caller_env(),
-             allow_list = FALSE) {
+             call = rlang::caller_env()) {
 
       legacy_cols <-
         tryCatch(
@@ -288,11 +287,6 @@ resolve_desc_vars <-
           ),
           error = function(...) NULL
         )
-
-      if (allow_list && is.list(legacy_cols)) {
-        legacy_cols <-
-          unlist(legacy_cols, recursive = TRUE, use.names = FALSE)
-      }
 
       if (is.character(legacy_cols)) {
         legacy_cols <- unname(legacy_cols)
