+++ R/desc_rch.R
@@ -66,8 +66,8 @@ desc_rch <- function(.data,
 ){
 
   check_data_link(.data)
-  drug_s <- resolve_desc_vars(.data, rlang::enquo(drug_s), col_arg = "drug_s", allow_list = TRUE)
-  adr_s <- resolve_desc_vars(.data, rlang::enquo(adr_s), col_arg = "adr_s", allow_list = TRUE)
+  drug_s <- resolve_desc_vars(.data, rlang::enquo(drug_s), col_arg = "drug_s")
+  adr_s <- resolve_desc_vars(.data, rlang::enquo(adr_s), col_arg = "adr_s")
 
   core_desc_rch <-
     function(one_drug,
