### 2026-06-25 2a3678c
switching tests to expect_snapshot
