### 2026-06-25 040fa22
Changes before error encountered
Agent-Logs-Url: https://github.com/pharmacologie-caen/vigicaen/sessions/776bb341-fda6-49c2-817e-1138d3cae4bc
