+++ tests/testthat/_snaps/get_drecno.md
@@ -13,7 +13,7 @@
       
       -- `d_sel`: Matching drugs --
       
-      -- v Matched drugs 
+      -- v Matched drugs --
       
       > `nivolumab`: "nivolumab"
       > `ipilimumab`: "ipilimumab"
@@ -38,7 +38,7 @@
       
       -- `d_sel`: Matching drugs --
       
-      -- v Matched drugs 
+      -- v Matched drugs --
       
       > `nivolumab`: "nivolumab" and "ipilimumab;nivolumab"
       > `ipilimumab`: "ipilimumab" and "ipilimumab;nivolumab"
@@ -59,7 +59,7 @@
       
       -- `d_sel`: Matching drugs --
       
-      -- v Matched drugs 
+      -- v Matched drugs --
       
       > `a`: "chlorphenamine;guaifenesin;paracetamol;pseudoephedrine;tipepidine", "cinnamomum spp.;ephedra spp.;glycyrrhiza spp.;paeonia spp.;paracetamol;pueraria spp.;zingiber officinale;ziziphus spp.", "eperisone;paracetamol", "guaifenesin;paracetamol;pseudoephedrine", "caffeine;chlorphenamine;guaifenesin;methylephedrine;paracetamol;phenylephrine", "ascorbic acid;chlorphenamine;cloperastine;coptis spp.;lysozyme;methylephedrine;paracetamol;platycodon grandiflorus", "cetirizine;paracetamol;phenylephrine;zinc", "chlorphenamine;cloperastine;lysozyme;paracetamol", "allobarbital;codeine;paracetamol;salicylamide", "atropine;codeine;papaverine;paracetamol", "codeine;paracetamol;pentobarbital", "amobarbital;caffeine;codeine;paracetamol", "amobarbital;caffeine;codeine;octibenzonium bromide;paracetamol", "amobarbital;caffeine;codeine;dihydroergotamine;paracetamol", "caffeine;codeine;meprobamate;paracetamol", "codeine;meprobamate;paracetamol", "caffeine;codeine;paracetamol;phenazone;prothipendyl", "codeine;methaqualone;paracetamol", ..., "atropa bella-donna;caffeine;dihydrocodeine;diphenylpyraline;guaifenesin;methylephedrine;noscapine;paracetamol", and "ascorbic acid;caffeine;chlorphenamine;dextromethorphan;guaifenesin;herbal nos;paracetamol;pseudoephedrine"
       
@@ -286,7 +286,7 @@
       
       -- `d_sel`: Matching drugs --
       
-      -- v Matched drugs 
+      -- v Matched drugs --
       
       > `medi`: "enalapril"
       > `mix`: "enalaprilat" and "enalapril"
@@ -328,7 +328,7 @@
       
       -- `d_sel`: Matching drugs --
       
-      -- v Matched drugs 
+      -- v Matched drugs --
       
       > `medi`: "enalapril", "enalapril;folic acid", and "captopril;enalapril"
       > `mix`: "enalaprilat", "enalapril", "enalapril;folic acid", and "captopril;enalapril"
@@ -362,7 +362,7 @@
       
       -- `d_sel`: Matching drugs --
       
-      -- v Matched drugs 
+      -- v Matched drugs --
       
       > `para`: "paracetamol"
       
@@ -380,7 +380,7 @@
       
       -- `d_sel`: Matching drugs --
       
-      -- v Matched drugs 
+      -- v Matched drugs --
       
       > `set1`: "enalapril"
       
@@ -399,7 +399,7 @@
       
       -- `d_sel`: Matching drugs --
       
-      -- v Matched drugs 
+      -- v Matched drugs --
       
[…]
