### 2026-06-25 a6e7b7f
style: align matched CLI headings across get_* helpers
