### 2026-06-26 a4ec632
style: refactor CLI output for lists and general messages
Updates output in get_* and tb_* functions to use cli::cli_li for list items
within cli::cli_ul blocks and cli::cli_text for other informational messages.
This ensures more semantically correct and consistent CLI formatting,
replacing previous uses of cli::cli_inform.
