+++ tests/testthat/_snaps/get_atc_code.md
@@ -1,3 +1,298 @@
+# verbose controls success messages
+
+    Code
+      get_atc_code(atc_sel = atc_sel, mp = mp_, thg_data = thg_, vigilyze = TRUE,
+        verbose = TRUE)
+    Message
+      
+      -- get_atc_code() --------------------------------------------------------------
+      
+      -- v Matched ATC classes (`atc_sel`) --
+      
+      * `l03_j01`: 77 DrecNo values
+      i vigilyze set to TRUE, extracting DrecNos (?get_atc_code for details)
+      i Set `verbose` to FALSE to suppress this section.
+      --------------------------------------------------------------------------------
+      
+    Output
+      $l03_j01
+       [1]   9557156  91158341  56557033 101174561  22001499 140973512  99724899
+       [8] 141246411 104122340  60718653  69765129 101027183   5258896  44647939
+      [15] 124637140  31247674 132648692  10139311  56417291  36181560  82788360
+      [22] 141692577  71743832  82311164  94074904  29537060  46518439  78649890
+      [29]  39166976 148385805 138750525  68390278 149023831  85622225 101111886
+      [36] 145488064  46977048  75455542  81115813 142084334  27681292  87226707
+      [43]  13273647  69587926 142536471   3869792  97730373  79877256  11808717
+      [50] 124496545 126197612   5938385  76444137  54287154  43444090  41756448
+      [57]  16359127  80471885  71156571  35160081  32023302 133641124  15218007
+      [64]  31302208  77713819  45992560  32542138  39319101  87207610 121794236
+      [71]  89998308  68274950  95345772  88621728  98412032  90280771  23656361
+      
+
+---
+
+    Code
+      get_atc_code(atc_sel = atc_sel, mp = mp_, thg_data = thg_, vigilyze = FALSE,
+        verbose = TRUE)
+    Message
+      -- get_atc_code() --------------------------------------------------------------
+      
+      -- v Matched ATC classes (`atc_sel`) --
+      
+      * `l03_j01`: 1721 Record_Id values
+      i vigilyze set to FALSE, extracting Record_Ids (?get_atc_code for details)
+      i Set `verbose` to FALSE to suppress this section.
+      --------------------------------------------------------------------------------
+    Output
+      $l03_j01
+         [1] 125147328  57248336  15618102 146672190  30845227   6195643   1112640
+         [8]  79914782  37956134 116958459  13498039  15910227  93398265 106221945
+        [15]  33621448  68043488 110623476 138070287  31167884  23958761  90045096
+        [22]  88137855 139483781  68982895 119554278  21845048 119910956 138206281
+        [29]  86235868 118980938  69953912  37116646 114268152 114695569 106324542
+        [36] 121980220   3281221  42760039  82024872  28583193  30759213 122994063
+        [43]  10982401 144271963  37717533 103074304  81782428 147634867 110229787
+        [50] 123114375    227991  76293770  19983283  54539470  39242206  51552919
+        [57]  80895520  65111524 108900152  15463762  58027079 123295297 123593888
+        [64] 127871362  98891689    558188 117457521  16559940 114583481 106863787
+        [71]  78076372 129874336   2267139  20361564 142960343 106389619 114792249
+        [78] 137669827  93031934   3170450  90777711  29245922  85456343 141028616
+        [85]  84914651 139328606 147585715  33464995 114390929  85636436  48662863
+        [92]  87373899 144645773  98605707 103553403  61364392 122375277  74658988
+        [99]  97447480  18834254 140926119  24144176  33260415  59579483  78711722
+       [106]  17055738  62066611 148391118  95236813  23722800  96202921 122885565
[…]
