### 2026-06-26 dcf06ff
Deploying to gh-pages from @ pharmacologie-caen/vigicaen@212d55a67be0373535e8f640438015c4beee2cbd 🚀
