### 2026-06-26 ad50dc8
Deploying to gh-pages from @ pharmacologie-caen/vigicaen@bf50ee2ebb8adf94fa902cfd367d81f634133361 🚀
