+++ articles/basic_workflow.md
@@ -228,14 +228,10 @@ get_drecno(
 #> 
 #> ── `d_sel`: Matching drugs ──
 #> 
-#> ── ✔ Matched drugs
+#> ── ✔ Matched drugs ──
 #> 
-#> → `nivolumab`: "nivolumab" and "ipilimumab;nivolumab"
-#> 
-#> 
+#> • `nivolumab`: "nivolumab" and "ipilimumab;nivolumab"
 #> ℹ Set `verbose` to FALSE to suppress this section.
-#> 
-#> 
 #> 
 #> ────────────────────────────────────────────────────────────────────────────────
 #> $nivolumab
@@ -385,7 +381,16 @@ atc_drecno <-
     mp = mp,
     thg_data = thg_
     )
+#> 
+#> ── get_atc_code() ──────────────────────────────────────────────────────────────
+#> 
+#> ── ✔ Matched ATC classes (`atc_sel`) ──
+#> 
+#> • `l03`: 13 DrecNo values
 #> ℹ vigilyze set to TRUE, extracting DrecNos (?get_atc_code for details)
+#> ℹ Set `verbose` to FALSE to suppress this section.
+#> ────────────────────────────────────────────────────────────────────────────────
+#> 
 ```
 
 The
@@ -517,15 +522,11 @@ d_drecno <-
 #> 
 #> ── `d_sel`: Matching drugs ──
 #> 
-#> ── ✔ Matched drugs
+#> ── ✔ Matched drugs ──
 #> 
-#> → `nivolumab`: "nivolumab" and "ipilimumab;nivolumab"
-#> → `pembrolizumab`: "pembrolizumab"
-#> 
-#> 
+#> • `nivolumab`: "nivolumab" and "ipilimumab;nivolumab"
+#> • `pembrolizumab`: "pembrolizumab"
 #> ℹ Set `verbose` to FALSE to suppress this section.
-#> 
-#> 
 #> 
 #> ────────────────────────────────────────────────────────────────────────────────
 
@@ -567,15 +568,11 @@ d_drecno <-
 #> 
 #> ── `d_sel`: Matching drugs ──
 #> 
-#> ── ✔ Matched drugs
+#> ── ✔ Matched drugs ──
 #> 
-#> → `analgesics`: "paracetamol" and "tramadol"
-#> → `ici`: "nivolumab" and "pembrolizumab"
-#> 
-#> 
+#> • `analgesics`: "paracetamol" and "tramadol"
+#> • `ici`: "nivolumab" and "pembrolizumab"
 #> ℹ Set `verbose` to FALSE to suppress this section.
-#> 
-#> 
 #> 
 #> ────────────────────────────────────────────────────────────────────────────────
 
@@ -682,12 +679,10 @@ a_llt <-
 #> 
 #> ── ✔ Matched reactions at `pt` level (number of codes) ──
 #> 
-#> → `a_colitis`: "Autoimmune colitis (1)", "Colitis (25)", "Colitis microscopic
-#>   (3)", "Diarrhoea (53)", "Diarrhoea haemorrhagic (8)", "Duodenitis (5)",
-#>   "Enteritis (8)", "Enterocolitis (4)", "Enterocolitis haemorrhagic (10)", and
-#>   "Ulcerative gastritis (1)"
-#> 
-#> 
+#> • `a_colitis`: "Autoimmune colitis (1)", "Colitis (25)", "Colitis microscopic
+#> (3)", "Diarrhoea (53)", "Diarrhoea haemorrhagic (8)", "Duodenitis (5)",
+#> "Enteritis (8)", "Enterocolitis (4)", "Enterocolitis haemorrhagic (10)", and
+#> "Ulcerative gastritis (1)"
 #> ℹ Set `verbose` to FALSE to suppress this section.
 
 a_llt
[…]
