### 2026-06-29 7c8f00c
docs: add full CeCILL 2.1 license text as build-ignored LICENSE.md
Restore the full OSI-approved CeCILL 2.1 license text to the repository
in LICENSE.md (byte-identical to the text removed in 439a705). The file
is already listed in .Rbuildignore (^LICENSE\.md$), so it is excluded
from the package build and does not re-introduce the CRAN "Non-FOSS
package license (file LICENSE)" note that prompted its earlier removal.

This keeps the full license text living with the code on GitHub, as
requested in the JOSS review (openjournals/joss-reviews#10321), while
preserving CRAN compliance. License fiel
[…]
