### 2026-06-29 a48406b
feat: tidy-select support in desc_facvar() and desc_cont()
desc_facvar() and desc_cont() now accept tidy-select syntax (bare names,
where(), starts_with(), all_of(), integer positions, ...) in addition to a
character vector of column names. Selections are resolved by a new internal
resolve_columns() helper; the legacy character-vector path is unchanged, so
existing behaviour and error classes (columns_not_in_data,
columns_not_numeric_integer) are preserved.

Also fixes desc_cont() to always return a plain data.frame, as documented.
Previously an in-memory Arrow Table produced a tibble, making 
[…]
