### 2026-06-29 1efb934
docs: add FAQ and frequent errors article
Adds a pkgdown article (vignettes/articles/faq.Rmd) collecting common
errors and their solutions, starting with the arrow 'malloc ... failed'
out-of-memory error and the vigi_routine() drug/effect name mismatch.

The article is registered under a new 'Help' section in _pkgdown.yml and
build-ignored (vignettes/articles is pkgdown-only, not a CRAN vignette),
so it can grow as new errors are reported.

Refs #159
