### 2026-06-29 dff4d18
refactor: extract vigi_routine() low-memory bypasses into tested helpers
vigi_routine() avoids OOM on low-spec machines (#158) by computing the
single drug-adr disproportionality from case-id set sizes and the 2x2
counts directly, instead of adding drug/adr columns to the full demo
table (which add_drug()/add_adr() would materialise in memory).

That bypass logic lived inline in the (graphics-producing, hard-to-test)
vigi_routine(). Extract it into two internal helpers:

* vr_case_sets()      - drug / adr / intersection UMCReportId sets
* vr_ic_from_counts() - Information Component from a 2x2 
[…]
