### 2026-06-29 331d20c
fix: validate repbasis, restore arrow option on exit, unify is_arrow (#183)
Three converged robustness/consistency fixes from the package assessment:

* parse_repbasis(): validate repbasis (^[sci]+$) and de-duplicate the
  identical grepl() block from add_drug()/add_dose()/vigi_routine(). An
  empty/invalid value (e.g. "" or "xyz") now errors instead of silently
  filtering the drug table to zero rows.

* add_dose()/add_ind()/vigi_routine() restore options(arrow.pull_as_vector)
  via on.exit(), so an error mid-function no longer leaks the user's global
  option (matching add_outcomes()). Captu
[…]
