+++ R/get_atc_code.R
@@ -71,14 +71,14 @@ get_atc_code <-
       cli::cli_warn("names of atc_sel were tolower-ed and trimed")
     }
 
-    if(any(c("Table", "Dataset") %in% class(mp))){
+    if(is_arrow(mp)){
       # automatically collect mp if out of memory
       # since it's a small table
       mp <-
         dplyr::collect(mp)
     }
 
-    if(any(c("Table", "Dataset") %in% class(thg_data))){
+    if(is_arrow(thg_data)){
       # automatically collect thg_data if out of memory
       # since it's a small table
       thg_data <-
