### 2026-06-29 0afb0fe
perf: stream screen_adr() aggregation, add column validation (#189)
screen_adr() coerced both inputs to data.table up front
(as.data.table(.data)), which fully materialised the adr table in RAM
when given an out-of-memory arrow Dataset and pulled every column. On the
full 107M-row adr that peaked at ~10.5 GB.

Rewrite the body in dplyr/arrow so the per-term aggregation streams and
only the small result is collected: peak dropped to ~2.8 GB on the same
data (fits an 8 GB machine, where the old path OOMs). Also:

* validate the columns actually used (UMCReportId/MedDRA_Id in .data,
  llt_code + 
[…]
