### 2026-06-29 ee59183
test: cover screen_adr() empty-input guard (#189)
Add a test exercising the empty-.data abort (class empty_data_no_reports),
restoring 100% coverage flagged by codecov on #190.
