### 2026-06-29 8c5bd0f
feat: add col_select argument to dt_parquet() (#193)
dt_parquet(in_memory = TRUE) read every column; there was no way to load
a subset, and on the big main tables that is the main RAM driver.

Add an optional tidy-select col_select (default NULL = all columns),
forwarded to arrow::read_parquet(col_select=) in memory and to
dplyr::select() for the out-of-memory open_dataset() path. Purely
additive - existing calls are unchanged.

Adds col_select tests (in-memory character + bare tidy-select, and
out-of-memory). dt_parquet() coverage 100%; full suite green.

Closes #193
