### 2026-07-27 cfd5331
Refactor summary section and statement of need
https://github.com/openjournals/joss-reviews/issues/10321#issuecomment-5079762309
