### 2026-07-27 d85d424
Remove most usage information and refer to vignettes
