+++ paper.md
@@ -109,12 +109,10 @@ The vigicaen package aims at providing a toolbox for the VigiBase®
 Extract Case Level database, tackling several technical challenges to run
 on low-specification computers, and providing easy and reproducible access
 to advanced features.[@dolladille2025] This article will explain the
-technical choices and data management logic underlying the package, and
-provide some examples of its main features. Additional examples and use
-cases are covered in the package vignettes, which can be found on the
-package website at <https://pharmacologie-caen.github.io/vigicaen/>. Of
-important note, the package is not supported by, nor does it reflect the opinion of,
-the WHO. The Uppsala Monitoring Centre, in charge of maintaining
+technical choices and computing logic of the package. Examples and use
+cases are covered in the package vignettes, available on the
+package website at <https://pharmacologie-caen.github.io/vigicaen/>. 
+The Uppsala Monitoring Centre, in charge of maintaining
 VigiBase®, was informed of the package development and kindly allowed
 its publication, acknowledging the potential benefit of promoting the use
 of VigiBase®.
@@ -202,48 +200,18 @@ development of specific functions like `vigi_routine`.
 
 # Processing `vigibase®` source files.
 
-Clinicians and pharmacovigilance researchers are used to working with
-low-specification computers. The typical available Random Access Memory
-rarely exceeds 16GB, which is one of the key bottlenecks when dealing with
-large data files in R.[@22arro] VigiBase® Extract Case Level files
+VigiBase® Extract Case Level files
 currently exceed 30GB once unpacked, which is too large to be loaded
-in-memory by mainstream readers like `read.table()`.
+in-memory by mainstream readers like `read.table()` on most computers. Vigicaen 
+relies on `parquet` files, a recent format based on open
+standards, supported by Arrow.[@parquet; @apachea; @arrowr] Datasets remain out of
+memory. Various tests of vigicaen on 16GB RAM computers succeeded in
+processing the source files.
 
-Vigicaen relies on `parquet` files, a recent format based on open
-standards.[@parquet] Arrow is a cross-language development platform that
-allows for the manipulation of large datasets.[@apachea] It is implemented
-in R via the arrow package.[@arrowr] Datasets remain out of
-memory, allowing for the processing of large files on low-specification
-computers. Various tests of vigicaen on 16GB RAM computers succeeded in
-processing the source files. This, in combination with an alignment as close as
-possible to the tidyverse style guide, is also aimed at
-providing a modern and more rigorous approach compared to base R.[@wickham2023]
-
-Sourcing VigiBase® Extract Case Level files is done with the `tb_*`
-family of functions.
-
-First, we define paths to the source folders.
+Below, we provide a partial example of the sourcing process, with `tb_vigibase()`.
+Users may refer to the vignettes for more details.
 
 
-``` r
-library(vigicaen)
-
-path_base <- paste0(tempdir(), "/main/")
-path_sub  <- paste0(tempdir(), "/sub/")
-
-dir.create(path_base)
-dir.create(path_sub)
-```
-
-Example files can be put in these folders.
-
-
-``` r
-create_ex_main_txt(path_base)
-create_ex_sub_txt(path_sub)
-```
-
-Then, we run the related `tb_*` function, `tb_vigibase()`.
 
 
 ``` r
@@ -254,143 +222,22 @@ tb_vigibase(path_base, path_sub)
 ```
 ## -- tb_vigibase() ---------------------------------------------------------------
 ## 
+## v All expected csv files found in `path_base` and `path_sub`
+## 
 ## i Checking for existing tables.
 ## 
 ## i Creating vigibase tables.
 ## 
 ## This process must only be done once per database version.
 ## It can take up to 30minutes.
[…]
