### 2026-07-27 d503cfb
Fix Minoc reference and call from paper.md
