+++ paper.bib
@@ -218,7 +218,9 @@ PMID: 40617160},
 	year = {2025},
 	month = {02},
 	date = {2025-02-04},
-	journal = {Therapie},
+	journal = {Therapies},
+	pages = {561--571},
+	volume = {80},
 	doi = {10.1016/j.therap.2025.01.004},
 	note = {PMID: 40037976},
 	langid = {eng}
