### 2026-07-27 8a92dce
Refactor to shorten manuscript
