### 2026-08-04 d2b00ef
Fix: remove utils global variables in screen_adr, blocking roxygen doc
Had to define said variables elsewhere.
