### 2026-08-04 7cd2c49
Deploying to gh-pages from @ pharmacologie-caen/vigicaen@d2b00ef44f0d96b22d0104422224a71867d6b0a3 🚀
