+++ reference/index.md
@@ -80,7 +80,7 @@
   : Sample of WHODrug
 - [`nice_p()`](https://pharmacologie-caen.github.io/vigicaen/reference/nice_p.md)
   **\[stable\]** : Nice printing of p-values
-- [`screen_adr`](https://pharmacologie-caen.github.io/vigicaen/reference/screen_adr.md)
+- [`screen_adr()`](https://pharmacologie-caen.github.io/vigicaen/reference/screen_adr.md)
   **\[stable\]** : Screening of adverse drug reactions
 - [`screen_drug()`](https://pharmacologie-caen.github.io/vigicaen/reference/screen_drug.md)
   **\[stable\]** : Screening of drugs
