+++ revdep/README.md
@@ -2,7 +2,7 @@
 
 |field    |value                                                                                |
 |:--------|:------------------------------------------------------------------------------------|
-|version  |R version 4.6.0 (2026-04-24 ucrt)                                                    |
+|version  |R version 4.6.1 (2026-06-24 ucrt)                                                    |
 |os       |Windows 11 x64 (build 26200)                                                         |
 |system   |x86_64, mingw32                                                                      |
 |ui       |RStudio                                                                              |
@@ -10,16 +10,16 @@
 |collate  |French_France.utf8                                                                   |
 |ctype    |French_France.utf8                                                                   |
 |tz       |Europe/Paris                                                                         |
-|date     |2026-06-24                                                                           |
-|rstudio  |2026.01.1+403 Apple Blossom (desktop)                                                |
-|pandoc   |3.6.3 @ C:/Program Files/RStudio/resources/app/bin/quarto/bin/tools/ (via rmarkdown) |
-|quarto   |1.8.25 @ C:\PROGRA~1\RStudio\RESOUR~1\app\bin\quarto\bin\quarto.exe                  |
+|date     |2026-08-07                                                                           |
+|rstudio  |2026.06.0+242 Blue Plumbago (desktop)                                                |
+|pandoc   |3.8.3 @ C:/Program Files/RStudio/resources/app/bin/quarto/bin/tools/ (via rmarkdown) |
+|quarto   |1.9.38 @ C:\PROGRA~1\RStudio\RESOUR~1\app\bin\quarto\bin\quarto.exe                  |
 
 # Dependencies
 
-|package  |old   |new   |Δ  |
-|:--------|:-----|:-----|:--|
-|vigicaen |1.0.0 |2.0.0 |*  |
+|package  |old   |new        |Δ  |
+|:--------|:-----|:----------|:--|
+|vigicaen |2.0.0 |2.0.0.9000 |*  |
 
 # Revdeps
 
