### 2026-08-07 6079dfb
Deploying to gh-pages from @ pharmacologie-caen/vigicaen@ac64f6b2ccabd8e9b43f3303b2aea84a73db9ae2 🚀
