### 2026-08-08 833ea73
Fix typos and references
Fixes for https://github.com/openjournals/joss-reviews/issues/10321#issuecomment-5151451658
