+++ paper.bib
@@ -129,13 +129,13 @@ PMCID: PMC6331976},
 	url = {https://cran.r-project.org/web/packages/arrow/index.html}
 }
 
-@book{wickham2023,
-	title = {tidyverse: Easily Install and Load the 'Tidyverse'},
-	author = {Wickham, Hadley and RStudio, },
-	year = {2023},
-	month = {02},
-	date = {2023-02-22},
-	url = {https://cran.r-project.org/web/packages/tidyverse/index.html}
+@Manual{wickham2023,
+  title = {tidyverse: Easily Install and Load the 'Tidyverse'},
+  author = {Wickham, Hadley and RStudio},
+  year = {2023},
+  note = {R package},
+  url = {https://CRAN.R-project.org/package=tidyverse},
+  doi = {10.32614/CRAN.package.tidyverse}
 }
 
 @article{legallois2025,
@@ -266,15 +266,11 @@ for Present},
     doi = {10.32614/CRAN.package.vaersvax},
 }
 
-@misc{embry_vaersndvax_2016,
-	title = {{vaersNDvax}: {Non}-{Domestic} {Vaccine} {Adverse} {Event} {Reporting} {System} ({VAERS}) {Vaccine} {Data} for {Present}},
-	copyright = {CC0},
-	shorttitle = {{vaersNDvax}},
-	url = {https://cran.r-project.org/web/packages/vaersNDvax/index.html},
-	abstract = {Non-Domestic VAERS vaccine data for 01/01/2016 - 06/14/2016. If you want to explore the full VAERS data for 1990 - Present (data, symptoms, and vaccines), then check out the 'vaersND' package from the URL below. The URL and BugReports below correspond to the 'vaersND' package, of which 'vaersNDvax' is a small subset (2016 only). 'vaersND' is not hosted on CRAN due to the large size of the data set. To install the Suggested 'vaers' and 'vaersND' packages, use the following R code: 'devtools::install\_git("https://gitlab.com/iembry/vaers.git", build\_vignettes = TRUE)' and 'devtools::install\_git("https://gitlab.com/iembry/vaersND.git", build\_vignettes = TRUE)'. "VAERS is a national vaccine safety surveillance program co-sponsored by the US Centers for Disease Control and Prevention (CDC) and the US Food and Drug Administration (FDA). VAERS is a post-marketing safety surveillance program, collecting information about adverse events (possible side effects) that occur after the administration of vaccines licensed for use in the United States." For more information about the data, visit {\textless}https://vaers.hhs.gov/index{\textgreater}. For information about vaccination/immunization hazards, visit {\textless}http://www.questionuniverse.com/rethink.html/\#vaccine{\textgreater}.},
-	urldate = {2026-04-27},
-	author = {Embry, Irucka},
-	month = aug,
-	year = {2016},
+@Manual{embry_vaersndvax_2016,
+  title = {vaersNDvax: Non-Domestic Vaccine Adverse Event Reporting System (VAERS) Vaccine Data for Present},
+  author = {Embry, Irucka},
+  year = {2016},
+  note = {R package},
+  url = {https://CRAN.R-project.org/package=vaersNDvax},
+  doi = {10.32614/CRAN.package.vaersNDvax}
 }
-
