### 2026-08-18 #10321 @rkurchin
👋 Hi, congrats on being nearly done! As AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks, some of which may require your attention:
- [ ] please remove verbatim text styling in title for both paper and archive
- [ ] please add space before citations
- [ ] second-to-last paragraph of statement of need: homemade is one (unhyphenated) word
- [ ] second-to-last bullet point under software design: suggest change to "provision of help" for better parallelism
