### 2026-08-25 ca21338
Fixes for proof checking
https://github.com/openjournals/joss-reviews/issues/10321#issuecomment-5322217439
