### 2026-08-25 #10321 @cdolladille
Dear @rkurchin, thank you for this proof checking of the article. My apologies for the delay.
In ca213385183db901fe6e6cc1acf378676654182c, the following changes were made

- [x] please remove verbatim text styling in title for both paper and archive -> Done in both
- [x] please add space before citations -> Added before each citation
- [x] second-to-last paragraph of statement of need: homemade is one (unhyphenated) word -> Removed the hyphen
- [x] second-to-last bullet point under software design: suggest change to "provision of help" for better parallelism -> Change applied

I remain at your entire disposal for further adjustments.
Charles
