### 2026-08-30 #10321 @rkurchin
Congrats to @cdolladille and thanks to handling editor @csoneson and reviewers @mhovd and @pacoramon. JOSS wouldn't exist without your volunteer efforts!
